Add extra christmas tree, info signpost, puppet show hangout zones
