+++ src/modules/events/views/ui/layouts/types/onsite/christmas-onsite-2020/rooms/ChristmasTree.tsx
@@ -1,0 +1,516 @@
+import { memo } from "react";
+import styled from "styled-components";
+
+import Room from "./Room";
+
+const StyledRoom = styled(Room)`
+    width: 18%;
+    position: absolute;
+    left: 40.5%;
+    top: -7vw;
+    z-index: 4;
+`;
+
+interface ChristmasTreeProps extends React.ComponentPropsWithoutRef<typeof Room> {}
+
+const ChristmasTree: React.FC<ChristmasTreeProps> = (props) => {
+    return (
+        <StyledRoom {...props}>
+            <svg viewBox='0 0 798.88 817.21'>
+                <defs>
+                    <style>
+                        {
+                            ".christmas_tree__cls-3{fill:#fcf8cd}.christmas_tree__cls-4{fill:#ffe09c}.christmas_tree__cls-5{fill:#fff}.christmas_tree__cls-6{fill:#ee6d55}.christmas_tree__cls-7{fill:#452a51}.christmas_tree__cls-11{fill:#000543}.christmas_tree__cls-12{fill:#ff9985}.christmas_tree__cls-16{fill:#fa694f}.christmas_tree__cls-17{fill:#cde9f9}.christmas_tree__cls-18{fill:#0b245e}.christmas_tree__cls-19{fill:#d15447}.christmas_tree__cls-20{fill:none}.christmas_tree__cls-21{fill:#498ac4}.christmas_tree__cls-22{fill:#141f49}"
+                        }
+                    </style>
+                </defs>
+                <g id='christmas_tree__Layer_2' data-name='Layer 2'>
+                    <g id='christmas_tree___Layer_' data-name='&lt;Layer&gt;'>
+                        <g id='christmas_tree__Christmas_tree' data-name='Christmas tree'>
+                            <path
+                                d='M694 715.3l-96.28-113.46c-13-15.28-2-38.63 18.12-38.63h4c20.14 0 31.09-23.35 18.12-38.64l-97.12-114.45c-13-15.28-2-38.63 18.12-38.63h23.57c20.14 0 31.09-23.35 18.12-38.64l-80.27-94.54c-13-15.29-2-38.64 18.12-38.64s31.1-23.35 18.12-38.63l-78.41-92.37a23.81 23.81 0 00-36.24 0L363.55 161c-13 15.28-2 38.63 18.12 38.63s31.1 23.35 18.12 38.64l-80.26 94.54c-13 15.28-2 38.64 18.12 38.64h23.57c20.14 0 31.09 23.35 18.12 38.63l-97.17 114.49c-13 15.29-2 38.64 18.12 38.64h4c20.14 0 31.09 23.35 18.12 38.63L226.14 715.3c-13 15.29-2 38.64 18.12 38.64h431.66c20.14 0 31.08-23.35 18.08-38.64z'
+                                fill='#4859a3'
+                            />
+                            <path fill='#2e377a' d='M437.35 753.27h45.55v63.94h-45.55z' />
+                            <ellipse className='christmas_tree__cls-3' cx={291.82} cy={522.59} rx={10.89} ry={10.81} />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M365.84 533.3a10.9 10.9 0 11-10.9-10.81 10.86 10.86 0 0110.9 10.81z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M430 528.63a10.9 10.9 0 11-10.89-10.81A10.85 10.85 0 01430 528.63z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M488 514.08a10.9 10.9 0 11-10.89-10.81A10.85 10.85 0 01488 514.08z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M544.11 496.5a10.9 10.9 0 11-10.9-10.81 10.86 10.86 0 0110.9 10.81z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M593.65 465.52a10.9 10.9 0 11-10.9-10.81 10.86 10.86 0 0110.9 10.81zM392.11 378.74a8.73 8.73 0 01-5.47 11.11 8.83 8.83 0 01-11.2-5.42 8.73 8.73 0 015.46-11.12 8.84 8.84 0 0111.21 5.43z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M427.4 382.29a8.71 8.71 0 01-5.47 11.11 8.83 8.83 0 01-11.2-5.43 8.73 8.73 0 015.47-11.11 8.83 8.83 0 0111.2 5.43z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M462.86 376a8.74 8.74 0 01-5.47 11.12 8.84 8.84 0 01-11.21-5.43 8.73 8.73 0 015.47-11.11 8.84 8.84 0 0111.21 5.42z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M499.49 363.37a8.73 8.73 0 01-5.49 11.11 8.85 8.85 0 01-11.21-5.43 8.73 8.73 0 015.47-11.11 8.84 8.84 0 0111.23 5.43z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M532.73 345.87a8.73 8.73 0 01-5.47 11.13 8.84 8.84 0 01-11.21-5.42 8.74 8.74 0 015.47-11.12 8.84 8.84 0 0111.21 5.41z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M561.26 325.59a8.72 8.72 0 01-5.47 11.12 8.84 8.84 0 01-11.2-5.43 8.72 8.72 0 015.47-11.12 8.83 8.83 0 0111.2 5.43z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M580.2 300.36a8.73 8.73 0 01-5.47 11.11 8.84 8.84 0 01-11.21-5.42 8.74 8.74 0 015.48-11.12 8.84 8.84 0 0111.2 5.43zM384.56 412.23a8.53 8.53 0 01-11.4 3.74 8.39 8.39 0 01-3.77-11.31 8.52 8.52 0 0111.4-3.75 8.4 8.4 0 013.77 11.32z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M403.86 433.77a8.53 8.53 0 01-11.4 3.74 8.39 8.39 0 01-3.77-11.31 8.52 8.52 0 0111.4-3.75 8.4 8.4 0 013.77 11.32z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M435.15 452.53a8.52 8.52 0 01-11.4 3.75A8.4 8.4 0 01420 445a8.54 8.54 0 0111.41-3.74 8.39 8.39 0 013.74 11.27z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M467.68 465.32a8.53 8.53 0 01-11.41 3.75 8.4 8.4 0 01-3.77-11.32 8.53 8.53 0 0111.4-3.75 8.39 8.39 0 013.78 11.32z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M504.13 469.3a8.53 8.53 0 01-11.41 3.75 8.4 8.4 0 01-3.72-11.32 8.54 8.54 0 0111.36-3.73 8.39 8.39 0 013.77 11.3z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M537.39 460.88a8.53 8.53 0 01-11.4 3.74 8.39 8.39 0 01-3.77-11.31 8.51 8.51 0 0111.4-3.74 8.39 8.39 0 013.77 11.31z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M567.89 444.11a8.53 8.53 0 01-11.4 3.74 8.39 8.39 0 01-3.77-11.31 8.52 8.52 0 0111.4-3.75 8.4 8.4 0 013.77 11.32z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M334.19 587.22a11.54 11.54 0 01-15.42 5.06 11.35 11.35 0 01-5.1-15.3 11.53 11.53 0 0115.42-5.06 11.36 11.36 0 015.1 15.3z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M376.69 617.72a11.53 11.53 0 01-15.42 5.06 11.34 11.34 0 01-5.1-15.29 11.53 11.53 0 0115.42-5.06 11.34 11.34 0 015.1 15.29z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M435.53 646.3a11.53 11.53 0 01-15.42 5.06 11.35 11.35 0 01-5.1-15.3 11.52 11.52 0 0115.42-5 11.34 11.34 0 015.1 15.24z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M496.74 657.69a11.52 11.52 0 01-15.42 5.06 11.34 11.34 0 01-5.1-15.29 11.53 11.53 0 0115.42-5.06 11.35 11.35 0 015.1 15.29z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M567.06 649.75a11.53 11.53 0 01-15.42 5.06 11.35 11.35 0 01-5.1-15.3 11.53 11.53 0 0115.46-5.06 11.35 11.35 0 015.06 15.3z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M628.42 630.6a11.53 11.53 0 01-15.42 5.06 11.34 11.34 0 01-5.1-15.29 11.53 11.53 0 0115.42-5.06 11.35 11.35 0 015.1 15.29z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M393.2 253.63a6.56 6.56 0 11-4.64-8 6.56 6.56 0 014.64 8z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M406.32 280.46a6.56 6.56 0 01-8 4.6 6.49 6.49 0 01-4.64-8 6.56 6.56 0 018-4.6 6.49 6.49 0 014.64 8z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M433.83 293.48a6.56 6.56 0 11-4.64-8 6.56 6.56 0 014.64 8z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M461.62 298.38a6.56 6.56 0 11-4.63-8 6.56 6.56 0 014.63 8z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M497.49 301a6.57 6.57 0 01-8 4.6 6.51 6.51 0 118-4.6z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M532 296.14a6.56 6.56 0 11-4.64-8 6.56 6.56 0 014.64 8z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M562.82 285a6.57 6.57 0 01-8 4.6 6.51 6.51 0 118-4.6z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M431 240.49a6.54 6.54 0 01-4.08 8.36 6.66 6.66 0 01-8.43-4.07 6.56 6.56 0 014.08-8.36 6.65 6.65 0 018.43 4.07z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M459.72 230.39a6.56 6.56 0 01-4.08 8.36 6.65 6.65 0 01-8.43-4.08 6.54 6.54 0 014.08-8.35 6.65 6.65 0 018.43 4.07z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M488.21 214.22a6.53 6.53 0 01-4.08 8.36 6.65 6.65 0 01-8.43-4.07 6.56 6.56 0 014.08-8.36 6.65 6.65 0 018.43 4.07z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M512.4 194.37a6.54 6.54 0 01-4.08 8.35 6.64 6.64 0 01-8.43-4.07 6.53 6.53 0 014.11-8.35 6.64 6.64 0 018.4 4.07z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M533.89 173.36a6.56 6.56 0 01-4.08 8.36 6.65 6.65 0 01-8.43-4.08 6.53 6.53 0 014.08-8.35 6.65 6.65 0 018.43 4.07z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M552.5 152.07a6.54 6.54 0 01-4.07 8.36 6.65 6.65 0 01-8.43-4.07 6.54 6.54 0 014.08-8.36 6.65 6.65 0 018.42 4.07zM410.22 112.78a5.23 5.23 0 11-2-7.11 5.2 5.2 0 012 7.11z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M425.22 129.68a5.23 5.23 0 11-2-7.12 5.21 5.21 0 012 7.12z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M446.3 141.37a5.23 5.23 0 11-2-7.11 5.2 5.2 0 012 7.11z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M470.84 149.72a5.23 5.23 0 11-2-7.11 5.2 5.2 0 012 7.11z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M495.32 153.6a5.23 5.23 0 11-2-7.12 5.21 5.21 0 012 7.12z'
+                            />
+                            <path
+                                className='christmas_tree__cls-4'
+                                d='M519.06 155.28a5.23 5.23 0 11-2-7.12 5.21 5.21 0 012 7.12z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M458.8 0l12.15 24.43 27.19 3.92-19.67 19.02 4.64 26.86-24.31-12.68-24.32 12.68 4.65-26.86-19.68-19.02 27.19-3.92L458.8 0z'
+                            />
+                        </g>
+                        <path
+                            className='christmas_tree__cls-5'
+                            d='M766.66 618.23c.44-.25.88-.5 1.33-.74a7.12 7.12 0 011.93-.74 6.09 6.09 0 011.18-.13 2.74 2.74 0 012.5 1.09c.32.64 0 1.6-.72 2.64a11.42 11.42 0 01-2.28 2.19l-2.12 1.61-3.94-4.68c.7-.42 1.4-.83 2.12-1.24z'
+                        />
+                        <path
+                            className='christmas_tree__cls-6'
+                            d='M739 629.42l-.16 2.68c-.08 1.29-.12 2.26-.13 3.26l3.42-2.54c8.25-6.13 14.21-10.55 21.1-14.66a2.62 2.62 0 013.33.56l3.45 4.09a3.33 3.33 0 01-.54 4.79c-13.48 10.28-20.89 16.74-25.32 20.6-5.47 4.77-7.31 6.38-10.86 6.38-5.56 0-9.74-4.33-11.18-11.62a70.69 70.69 0 01-1.54-11.3l-.09-2.76z'
+                        />
+                        <path
+                            className='christmas_tree__cls-5'
+                            d='M743 684.43a12.26 12.26 0 011.68 3.45c1 4.15-1.89 6.33-6 1.06-.89-1.14-1.76-2.23-2.61-3.29l5.82-3.07z'
+                        />
+                        <path
+                            className='christmas_tree__cls-6'
+                            d='M715.59 629.09l.19 1.9c1.36 14 4.88 18.44 11.88 27.27a196.56 196.56 0 0114.93 21l1.45 2.36-10.18 5.38-1.35-1.68c-6.77-8.45-12.41-14.76-16.94-19.83-7.88-8.83-12.63-14.15-15.15-21.26a87.28 87.28 0 01-3-15.47l-.37-3.55z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M696.62 595s30 5.37 29.67 4.22l16.07 3.06 3.4-.55c0 6.07-3.25 16.92-4.64 27.84a3.18 3.18 0 01-3.36 2.54l-36.49-1a4 4 0 01-4.07-3.39c-1.2-12.87-1.76-27.42-.58-32.72z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M699.4 593.68c2.77-8.89 9.35-15.65 12.75-24.69 1.63-4.31 3.27-14.52 3.41-18.92a33.84 33.84 0 010-3.78 8 8 0 00-1.36-4.91l8.21-12 2 1.28c9-3.8 15.78-.28 20.17 3.85 12.88 5.49 2.42 19.76 2.42 19.76-2.3 4.75-5.49 9.07-6.89 14.19a20.06 20.06 0 00-.65 5.08 21.64 21.64 0 00.33 2.64c1.18 6.15 3.27 11.34 3.71 17.56.23 3.31 1.39 5.57-.66 8.4-3.31 2-12.44.56-24.77-2.49-4.85-1.2-11.3 2.92-17.16-1.44a4.16 4.16 0 01-1.51-4.53z'
+                        />
+                        <path d='M720.31 532.2l.26-.15-.64.94a7.39 7.39 0 01.38-.79z' fill='#ff70ae' />
+                        <path
+                            d='M720.57 532.05c15-8.95 29.6 3.35 29.6 3.35-33.5 10.46-31.47.57-30.24-2.41z'
+                            fill='#e4e2fd'
+                        />
+                        <path
+                            d='M725.47 530.15a16.35 16.35 0 016.06-1.31c1.75-.08 5.08-.2 6.34 1.24s.7 4.23-.37 5.7a11.72 11.72 0 01-5.1 3.54 16.13 16.13 0 01-6.33 1.53 6.87 6.87 0 01-3.84-1 4.19 4.19 0 01-1.94-3.34 4.69 4.69 0 01.62-2.24 9.51 9.51 0 014.56-4.12z'
+                            fill='#ff9084'
+                        />
+                        <path
+                            className='christmas_tree__cls-5'
+                            d='M739.28 687.32l-6.28 2.39c-1.53.59-1.81 1.79-1.2 2.69l14.48-.41-.59-3.58a1.23 1.23 0 00-1.18-1.1z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M746.31 692l.16.6c.1.38-.3.73-.84.74l-11.26.16a2.92 2.92 0 01-2.54-1.09zM744.36 691.61h-1.82a.49.49 0 01-.5-.49c0-.42.22-.65.5-.65h1.82a.51.51 0 01.5.49c0 .42-.23.65-.5.65z'
+                        />
+                        <path
+                            className='christmas_tree__cls-5'
+                            d='M772.69 621.11l6.25 2.41c1.52.59 2.55-.1 2.69-1.17l-11-9.41-2 3a1.24 1.24 0 00.13 1.61z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M770.64 612.94l.29-.56c.18-.35.71-.34 1.12 0l8.44 7.43a2.91 2.91 0 011.14 2.52zM771.83 614.53l1.35 1.21a.5.5 0 010 .71c-.28.31-.6.32-.8.14l-1.35-1.22a.5.5 0 010-.7c.28-.31.59-.33.8-.14z'
+                        />
+                        <path
+                            className='christmas_tree__cls-12'
+                            d='M680.25 481a1.4 1.4 0 00-1.85.54c-.23.22-.5 1.66 1.21 5.17.15.29.38.73.68 1.26l4.06-2.85c-1.43-1.71-2.86-3.46-4.1-4.12z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M712.49 517.23c-.49-.91-8.88-11.38-9-11.47-2.56-3.05-5.44-6.77-8.23-10.38-2.46-3.18-4.78-6.18-6.64-8.39l-1-1.29a2.53 2.53 0 00-3.44-.55l-4.23 3a2.48 2.48 0 00-.74 3.24c3.64 6.5 17.77 29 23 36.3l2.35 3.16 8.73-12.18z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M682.72 476.48a.7.7 0 00-.8.56l-.19 1.07-.14-.16a.58.58 0 00-.11-.24l-3.15-4.12a.64.64 0 10-1 .78l2.15 2.81-.48.11-2.35-3.06a.64.64 0 00-1 .77l2.14 2.8c-.19.1-.37.22-.56.33l-1.68-2.13a.64.64 0 00-1 .77l1.64 2.14a1.07 1.07 0 00-.13.19l-.89-1.17a.65.65 0 00-.9-.12.64.64 0 00-.12.89l1.79 2.34a9.6 9.6 0 002 3.27c1.58 2.14 3 1.74 4 .08a5.16 5.16 0 00.66-3.64c0-.08.64-3.44.64-3.44a.69.69 0 00-.52-.83z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M709 515.74c1.27-.83 2.27-.28 3.31 1.11 3.82 5.09 12.46 15.13 12.46 15.13l-6 5.78c-2.76 2.66-1.74 9.84-4.39 6.32l-13.87-18.69a3.06 3.06 0 01.45-4.45z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M686.17 504.43q5 8.46 10.07 16.91c2.76 4.64 5.52 9.29 8.1 14a22.34 22.34 0 003.09 4.76c1.07 1.14 2.39 2 3.35 3.28 3 3.94 1.44 9.56 1.82 14.51.23 2.91 1.19 5.77 1 8.68-.2 3.35-1.9 6.46-2.24 9.8a16.13 16.13 0 01-.46 3.55c-.33 1-1 1.88-1.36 2.88a5.66 5.66 0 002.26 6.6 2 2 0 001.19.37c.7-.06 1.19-.7 1.57-1.3 3.42-5.44 4.88-11.91 5.44-18.31a59.93 59.93 0 00-.81-17.12c-2-9.8-6.84-18.78-12.3-27.16s-11.58-16.34-16.51-25c-2.28-4-6.66-9.91-11.34-11.61-1.47 5.36 5.31 12.11 7.13 15.16z'
+                        />
+                        <path
+                            className='christmas_tree__cls-12'
+                            d='M781.71 489.31a1.4 1.4 0 011.83.59c.22.24.45 1.68-1.37 5.14-.15.28-.4.71-.72 1.24l-4-3c1.55-1.68 3-3.39 4.26-3.97z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M748.37 524.49c.52-.89 9.23-11.1 9.31-11.18 2.65-3 5.64-6.61 8.54-10.13 2.56-3.1 5-6 6.9-8.19.35-.39.71-.8 1.09-1.25a2.53 2.53 0 013.46-.44l4.13 3.1a2.48 2.48 0 01.64 3.26c-3.83 6.4-18.65 28.4-24.09 35.59l-2.44 3.09-8.35-12.45z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M779.38 484.68a.7.7 0 01.78.58l.16 1.08.14-.16a.6.6 0 01.12-.24l3.27-4a.64.64 0 111 .8l-2.24 2.75a3.74 3.74 0 01.48.12l2.43-3a.64.64 0 011 .8l-2.23 2.73c.19.11.37.23.55.35l1.72-2.12a.65.65 0 01.9-.09.64.64 0 01.09.9l-1.7 2.09a1.22 1.22 0 01.12.19l.93-1.14a.64.64 0 01.9-.1.65.65 0 01.09.9l-1.89 2.29a9.66 9.66 0 01-2.14 3.21c-1.65 2.09-3.09 1.65-4-.05a5.18 5.18 0 01-.55-3.66c0-.08-.53-3.45-.53-3.45a.69.69 0 01.6-.78z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M751.94 523.11c-1.25-.86-2.26-.34-3.35 1-4 5-12.91 14.75-12.91 14.75l5.8 6c2.68 2.75 1.44 9.89 4.2 6.46L760.12 533a3.07 3.07 0 00-.32-4.46z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M747.38 555.29a9.15 9.15 0 011.9-4.56c1-1.21 2.34-2.07 3.44-3.18a22.26 22.26 0 003.24-4.67c2.72-4.66 5.62-9.22 8.53-13.77l10.58-16.6c1.91-3 8.9-9.54 7.58-14.93-4.72 1.55-9.29 7.3-11.69 11.25-5.19 8.54-11.56 16.32-17.27 24.53-4.72 6.77-9 14-11.68 21.75z'
+                        />
+                        <path
+                            className='christmas_tree__cls-7'
+                            d='M750.83 556.58c-1.74 3-4 5.73-4.8 9.08-1.36 5.94 2.38 11.87 2.45 18 .06 5.1-2.46 10-2.1 15.12.23 3.3 1.68 6.51 1.28 9.79-.52 4.31-4.11 7.79-4.62 12.1-.29 2.55.48 5.37-.88 7.54a8.74 8.74 0 01-2.45 2.32 2.19 2.19 0 01-1.82.6 2.09 2.09 0 01-1.09-1.34c-5.82-14-6.68-29.46-7.46-44.58-.23-4.6-.47-9.2-.34-13.8.06-2.24.26-4.62 1.62-6.4 1.08-1.42 2.74-2.24 4.19-3.27 8.43-5.94 9.76-18.15 17-25.48.9-.91 2.3-1.77 3.35-1a2.62 2.62 0 01.91 1.59c1.62 6.02-2.2 14.54-5.24 19.73zM705.41 543.26c.25 2.57.09 5.18.6 7.71.64 3.17 2.33 6.21 2 9.43-.36 4-3.63 7.13-4.56 11a51.77 51.77 0 00-.57 5.57c-.95 7.71-7.46 14.07-7.68 21.84-.1 3.44 1.06 6.87.64 10.28-.23 1.82-.9 3.56-1.29 5.35a17.8 17.8 0 003.2 14.36c.79 1 1.89 2.07 3.19 2 1.83-.12 2.76-2.22 3.15-4 .89-4.1 1.08-8.32 1.7-12.47 1.6-10.7 6-20.77 8.47-31.31 4.63-19.87 1.8-41.73-9.39-58.81-1.42-2.17-3.25-4.44-5.82-4.82-1.18 8.73 5.52 15.29 6.36 23.87z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M727.24 516.09c1.27-.43 3.06-1.93 4.39-2.09.5-.06 1 0 1.51-.1.21 0 1-.18 1-.43s-.1-.22-.25-.27a24 24 0 004.32-3.14 10.81 10.81 0 006.9-8.34 2.24 2.24 0 00.07-.34 1.74 1.74 0 00-.15-1 11.29 11.29 0 00-.88-1.21 6.08 6.08 0 01-.63-2.51 5.12 5.12 0 00-.25-1.08 4.15 4.15 0 012.36.23c2.49 1.32 5.39 3.61 8.11 4.38 6.56 1.87 11.48-3.78 12.77-5.12 7.88-8.17 8.7-13.07 8.75-24.4a18.05 18.05 0 00-1.27-7c-3.1-7.43-12.92-10.91-20.31-7.72l-.24.1c-7.42 3.37-11.38 12.56-9.37 20.44a18.37 18.37 0 011 4.55 3.86 3.86 0 01-2.26 3.74c-1.17.39-2.52-.05-3.65.44-2.84 1.23-1.28 6.2-2.81 8.53-.59-.29-1.16-.65-1.74-1a10.59 10.59 0 00-6-1.46 9.26 9.26 0 00-.92.13 6.9 6.9 0 00-3.82 2.22 5 5 0 00-.43.6 4.45 4.45 0 00-2.39 1.14 6.51 6.51 0 00-1.23 1.49 27.16 27.16 0 00-1.83 4.69 20.09 20.09 0 00-1 3 4.05 4.05 0 00.56 3.31 16.82 16.82 0 011.13 1.27 6.68 6.68 0 01.57 1.84c.44 1.74.4 4.06 2 4.87a8.81 8.81 0 005.99.24z'
+                        />
+                        <path
+                            d='M716.21 530.18c0 7.38 6.49 13.34 14.47 13.33s11.83-6 11.81-13.38 1.88-13.36-6.1-13.34-20.2 6.02-20.18 13.39z'
+                            fill='#514575'
+                        />
+                        <path
+                            d='M730.07 522.73h1.51a5.1 5.1 0 015.1 5.1v4.54a5.1 5.1 0 01-5.1 5.1h-1.51a5.1 5.1 0 01-5.1-5.1v-4.54a5.11 5.11 0 015.11-5.11z'
+                            transform='rotate(179.89 730.842 530.082)'
+                            fill='#fbe7d8'
+                        />
+                        <path
+                            d='M723.69 503.28l8.82 2.07a.75.75 0 00.89-.55l1.1-4.69a1.45 1.45 0 00-1.08-1.75l-8.12-1.9a3.5 3.5 0 10-1.61 6.82z'
+                            fill='#ffa27f'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M734.35 515.64a9.35 9.35 0 01-2.52 2.74 6.51 6.51 0 01-.79-4.89 3.23 3.23 0 001.54-2.27c.34-1.49-.07-2.86-.93-3.05a.47.47 0 00-.2 0l7-3.81c.04 2.8-1.03 6.47-4.1 11.28z'
+                        />
+                        <path
+                            className='christmas_tree__cls-12'
+                            d='M723.15 496.55a8.19 8.19 0 00-4.54 5.55l-1.48 6.44a8.26 8.26 0 0016.1 3.69l1.44-6.23z'
+                        />
+                        <path
+                            className='christmas_tree__cls-12'
+                            d='M732.59 531l-4.78.38a1.42 1.42 0 01-1.32-1.44L725.21 514a1.4 1.4 0 011.07-1.63l4.78-.39a1.42 1.42 0 011.32 1.45l1.28 15.89a1.41 1.41 0 01-1.07 1.68z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M726.36 503.16l2 .46a.54.54 0 01.4.64.55.55 0 01-.64.4l-2-.46a.54.54 0 01-.4-.64.53.53 0 01.64-.4zM719.21 501.48l1.52.36a.48.48 0 01.28.61.49.49 0 01-.53.43l-1.51-.36a.49.49 0 01-.29-.61.5.5 0 01.53-.43z'
+                        />
+                        <path
+                            className='christmas_tree__cls-16'
+                            d='M720.41 508.33l.1-.16a1 1 0 00.43 1l.81.49a.66.66 0 00.81-.15c0 .72-.63 1.12-1.09.84l-.81-.48a1.12 1.12 0 01-.25-1.54z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M725.88 506.55c-.12.49 0 .95.37 1s.68-.27.8-.76 0-1-.38-1-.67.26-.79.76z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M726.67 505.79l-.99-.24.47.76.52-.52zM718.86 504.9c-.12.49 0 1 .37 1s.68-.26.79-.76 0-1-.37-1-.65.26-.79.76z'
+                        />
+                        <path className='christmas_tree__cls-11' d='M719.65 504.14l-.99-.23.47.75.52-.52z' />
+                        <path
+                            className='christmas_tree__cls-12'
+                            d='M733.86 508.83a2.86 2.86 0 11-2.91 2.22 2.61 2.61 0 012.91-2.22z'
+                        />
+                        <path
+                            className='christmas_tree__cls-11'
+                            d='M737.17 503.43c-.47-2.07-.18-4.48-1.57-6.09a5.8 5.8 0 00-3.47-1.66 29.6 29.6 0 00-4.71-.34l-3.44-.07a.82.82 0 00-.38.06.78.78 0 00-.3.32c-.32.47-1.37 1.56-1.35 2.14s.78.84 1.28 1a12.62 12.62 0 018.29 7.36c.61 1.67.72 3.49 1.37 5.15a8.33 8.33 0 001.9 2.85c.11.09 4.47-5.86 4.47-5.86a23.15 23.15 0 01-2.09-4.86z'
+                        />
+                        <path
+                            className='christmas_tree__cls-5'
+                            d='M724.33 512.6c-.2.53-.39 1.06-1.62.94 0 0-2.07-.18-1.57-1.43z'
+                        />
+                        <path
+                            className='christmas_tree__cls-17'
+                            d='M707.45 467.38a2.85 2.85 0 102.84-2.85 2.85 2.85 0 00-2.84 2.85zM779.79 424.9a2.85 2.85 0 102.84-2.85 2.85 2.85 0 00-2.84 2.85zM658.84 443.07a1.64 1.64 0 101.63-1.64 1.63 1.63 0 00-1.63 1.64zM754.64 446.87a1.64 1.64 0 101.63-1.64 1.64 1.64 0 00-1.63 1.64zM656.59 463.57a3.93 3.93 0 103.92-3.93 3.92 3.92 0 00-3.92 3.93zM791 464.36a3.93 3.93 0 103.92-3.92 3.92 3.92 0 00-3.92 3.92zM710.86 453.31a4.4 4.4 0 104.39-4.39 4.39 4.39 0 00-4.39 4.39zM671.47 455.58a2.2 2.2 0 102.2-2.19 2.21 2.21 0 00-2.2 2.19zM731.34 468.41a2.2 2.2 0 102.2-2.2 2.19 2.19 0 00-2.2 2.2z'
+                        />
+                        <g id='christmas_tree__santa_claus' data-name='santa claus'>
+                            <path
+                                className='christmas_tree__cls-18'
+                                d='M131.91 792a25.28 25.28 0 00.11 2.68 2.16 2.16 0 00-.87 1.54l-.66 5.71 21.85-.44c.86-1.45.35-3.33-2-4.15l-9.55-3.34c0-1.1.07-2.31.14-3.36l-9 .23c0 .38-.02.75-.02 1.13zm4.51 9l-3.28.15a.5.5 0 01-.5-.51v-.73a.52.52 0 01.46-.56l3.28-.16a.52.52 0 01.51.52v.73a.5.5 0 01-.47.56z'
+                            />
+                            <path
+                                className='christmas_tree__cls-19'
+                                d='M145.48 745.16c-.39-25.49-1.75-48.54-3.52-66.35a8.12 8.12 0 01-3.89 5.61 15.9 15.9 0 01-7 2.08l-3.36.1-15.27.47c.94 3.61 3.45 8.62 3.76 10.69.4 2.73.37 5.51.7 8.25.81 6.64 3.74 13 3.69 19.67 0 2.31-.39 4.63-.05 6.92.5 3.28 2.44 6.19 3.22 9.42 1 4.28 0 8.74-.17 13.14a54.6 54.6 0 00.95 10.81 97.49 97.49 0 011.63 12.36 47.2 47.2 0 00.43 6.85 9.75 9.75 0 001.52 3.85 3.06 3.06 0 001.12 1 4.28 4.28 0 002.65.85l9.09-.23h.13a4.28 4.28 0 004.17-4.14c.15-5.38.25-10.73.31-16 .04-8.66.02-17.13-.11-25.35z'
+                            />
+                            <path
+                                className='christmas_tree__cls-11'
+                                d='M130.49 801.92l-.2 1c-.13.6.5 1.13 1.32 1.1l17-.58a4.42 4.42 0 003.77-1.93l-21.85.44zM136.35 799.22l-3.28.16a.52.52 0 00-.46.56v.73a.5.5 0 00.5.51l3.28-.15a.5.5 0 00.46-.56v-.73a.52.52 0 00-.5-.52z'
+                            />
+                            <path
+                                className='christmas_tree__cls-18'
+                                d='M85.77 790.87l-9-.23c.08 1.05.17 2.26.15 3.36l-9.57 3.33c-2.34.82-2.84 2.7-2 4.15l21.84.44-.66-5.71a2.1 2.1 0 00-.87-1.54 25.17 25.17 0 00.12-2.68c.02-.37.01-.74-.01-1.12zm-4.94 9.6v-.73a.51.51 0 01.5-.52l3.28.16a.52.52 0 01.47.56v.73a.5.5 0 01-.51.51l-3.28-.18a.52.52 0 01-.46-.53z'
+                            />
+                            <path
+                                className='christmas_tree__cls-19'
+                                d='M72.23 745.16c.4-25.49 1.76-48.54 3.53-66.35a8.09 8.09 0 003.89 5.61 15.9 15.9 0 007 2.08l3.35.1 15.27.47c-.94 3.61-3.45 8.62-3.75 10.69-.4 2.73-.37 5.51-.71 8.25-.81 6.64-3.74 13-3.69 19.67 0 2.31.4 4.63 0 6.92-.49 3.28-2.43 6.19-3.22 9.42-1 4.28 0 8.74.18 13.14a54.61 54.61 0 01-1 10.81 99.33 99.33 0 00-1.63 12.36 45.83 45.83 0 01-.42 6.85 9.75 9.75 0 01-1.38 3.82 3.17 3.17 0 01-1.13 1 4.26 4.26 0 01-2.65.85l-9.09-.23h-.12a4.27 4.27 0 01-4.17-4.14c-.15-5.38-.26-10.73-.31-16-.1-8.63-.07-17.1.05-25.32z'
+                            />
+                            <path
+                                className='christmas_tree__cls-11'
+                                d='M87.22 801.92l.21 1c.12.6-.51 1.13-1.32 1.1l-17-.58a4.38 4.38 0 01-3.76-1.93l21.84.44zM81.36 799.22l3.28.16a.52.52 0 01.47.56v.73a.5.5 0 01-.51.51l-3.31-.18a.52.52 0 01-.46-.56v-.73a.51.51 0 01.53-.49z'
+                            />
+                            <path
+                                className='christmas_tree__cls-20'
+                                d='M151.93 664.58c.18.82.34 1.65.49 2.48 1.23-3.78 2.89-7.42 3.9-11.27.34-1.27 1.1-3.47.83-4.83-.34-1.64-1.64-1.92-3.17-2.77a84.879 84.879 0 00-7.2-3.61c.73 7.01 3.64 13.19 5.15 20z'
+                            />
+                            <path
+                                className='christmas_tree__cls-21'
+                                d='M121.25 615.77a60.54 60.54 0 001.75 6.33c.86-1.16 1.73-2.3 2.61-3.44-1-2.59-2.35-5.63-4.48-7.16l-.82-.07.29.74a25 25 0 01.65 3.6zM132.91 636.87c-.55-1.69-1.17-3.37-1.83-5l-1.86 2.44a9.57 9.57 0 01-.83 1c2.08 4.39 4.34 8.74 6.49 13.12-.46-3.96-.77-7.83-1.97-11.56z'
+                            />
+                            <path
+                                className='christmas_tree__cls-21'
+                                d='M125.61 618.66c1.24-1.6 2.49-3.18 3.8-4.7a17.15 17.15 0 011.52-1.55l-1.18-.11-5.74-.53-2.88-.27c2.13 1.5 3.46 4.57 4.48 7.16z'
+                            />
+                            <path
+                                className='christmas_tree__cls-6'
+                                d='M184.3 646.13c-1.58-9.8-9.45-20.3-16.57-26.93-5.55-5.18-12.1-9.33-18.7-13.3 0-.12-9.77-5.85-9.77-5.85a5 5 0 00-2.14-.65c-6.07-.47-30.13-.25-30.13-.25-9.1-.11-18.14-.57-27.36 1.37C69 602.8 60.61 611 57.31 621.32L55 628.57a4.42 4.42 0 01-5.74 2.32l-25.72-10.5a5.27 5.27 0 00-6.85 2.87l-1.94 4.79a5.21 5.21 0 002.14 6.43l32.38 19.6a14.65 14.65 0 0017.43-1.67l6.45-5.75a43.24 43.24 0 01-.72 13.91c-5.29 10.26-8.28 22.8-6.33 34.42-.59 3-1.29 5.94-2.06 8.88-.95 3.69-1.63 7.38 1.51 10.28s8.25 3.18 12.39 3.51c5 .39 10 .64 15 .9 20.94 1.14 41.87 1.19 62.58-2.46 4-.71 5.53-5.89 3.69-9.08A72.68 72.68 0 01153 693a36.61 36.61 0 00.66-5.44c8.09 5.33 14.8-5.19 19.7-11.69 6.32-8.32 12.69-18.87 10.94-29.74zm-28 9.66c-1 3.85-2.68 7.49-3.91 11.26-.15-.82-.31-1.64-.5-2.47-1.5-6.81-4.41-13-5.14-20 2.46 1.09 4.85 2.29 7.2 3.6 1.53.86 2.83 1.14 3.16 2.78.31 1.36-.45 3.57-.78 4.83z'
+                            />
+                            <path
+                                className='christmas_tree__cls-22'
+                                d='M67 675.2l85.72-1 1.3 5.14a18.62 18.62 0 01.13 8.56L153 693l-87.23-.26a33.93 33.93 0 01-.47-10.08v-.26z'
+                            />
+                            <path
+                                className='christmas_tree__cls-12'
+                                d='M9.62 624.92a6.56 6.56 0 003-4.09l2.91 1.69q2.16 1.26 4.33 2.51l-3.17 5.35q-2-1.61-4.05-3.15c-1.06-.77-2.07-1.59-3.02-2.31zM10.47 613.4a1 1 0 00-.58 1.33l1.79 4.87a1 1 0 101.87-.68l-1.79-4.87a1 1 0 00-1.29-.65z'
+                            />
+                            <path
+                                className='christmas_tree__cls-12'
+                                d='M5.38 624.29a4.37 4.37 0 01-1.24-2 2.55 2.55 0 01.26-1.45 11.81 11.81 0 012-3.18 3.48 3.48 0 013.47-1.39 6 6 0 013.57 2.57c1.49 2.25 2 3.07 2 5.29 0 2.77-1.54 4.32-5.07 2.81-1.84-.79-3.73-1.29-4.99-2.65z'
+                            />
+                            <path
+                                className='christmas_tree__cls-12'
+                                d='M9.24 617.88L2.31 615a.92.92 0 01.7-1.71l6.93 2.85a.92.92 0 11-.7 1.7zM7.59 619.91l-6.92-2.85a.92.92 0 01.7-1.71l6.93 2.85a.93.93 0 11-.71 1.71zM7.5 622.86L.57 620a.92.92 0 11.71-1.7l6.92 2.85a.92.92 0 11-.7 1.7zM8.41 625.56l-6.93-2.85a.92.92 0 11.71-1.7l6.92 2.85a.92.92 0 01-.7 1.7zM147.39 688.24c-.23-1.87-.79-2.83-2.4-4.43l2.79-1.87 4.15-2.79 3.43 5.2c-1.55.75-3.09 1.5-4.6 2.25-1.13.56-2.31 1.11-3.37 1.64zM137.39 682.43a1 1 0 00.94 1.09l5.17.54a1 1 0 001.14-.88 1 1 0 00-.94-1.1l-5.17-.53a1 1 0 00-1.14.88z'
+                            />
+                            <path
+                                className='christmas_tree__cls-12'
+                                d='M145 691.78a4.41 4.41 0 01-2.34.23 2.65 2.65 0 01-1.2-.87 12 12 0 01-2-3.17 3.51 3.51 0 01.27-3.73 5.87 5.87 0 013.87-2.07c2.67-.35 3.63-.45 5.63.52 2.5 1.23 3.23 3.28.33 5.79-1.56 1.29-2.81 2.76-4.56 3.3z'
+                            />
+                            <path
+                                className='christmas_tree__cls-12'
+                                d='M140.89 685.5l-5.59 5a.92.92 0 11-1.22-1.38l5.58-5a.92.92 0 011.23 1.38zM142 687.86l-5.58 5a.93.93 0 01-1.3-.08.92.92 0 01.07-1.3l5.59-5a.92.92 0 011.22 1.38zM144.61 689.25l-5.58 5a.92.92 0 01-1.3-.07.94.94 0 01.07-1.31l5.59-5a.92.92 0 011.22 1.38zM147.44 689.62l-5.59 5a.92.92 0 11-1.22-1.38l5.58-5a.92.92 0 011.3.07.93.93 0 01-.07 1.31z'
+                            />
+                            <path
+                                className='christmas_tree__cls-6'
+                                d='M159.45 683a15.88 15.88 0 01-1.55-9.13c.83-4.53-6.12-6.47-7-1.91-.95 5.19.17 9.91 2.27 14.67.8 1.8 3.38 2.18 4.93 1.3 1.85-1.05 2.08-3.16 1.3-4.93z'
+                            />
+                            <path
+                                className='christmas_tree__cls-20'
+                                d='M110.54 680.5h9.73v-3.69a1.61 1.61 0 00-1.61-1.61h-16.24a1.62 1.62 0 00-1.62 1.61v13.36a1.62 1.62 0 001.62 1.61h16.24a1.61 1.61 0 001.61-1.61v-4.42h-9.73z'
+                            />
+                            <path
+                                className='christmas_tree__cls-3'
+                                d='M118.66 671.08h-16.24a5.74 5.74 0 00-5.73 5.73v13.36a5.74 5.74 0 005.73 5.73h16.24a5.74 5.74 0 005.73-5.73v-4.42h-4.12v4.42a1.61 1.61 0 01-1.61 1.61h-16.24a1.62 1.62 0 01-1.62-1.61v-13.36a1.62 1.62 0 011.62-1.61h16.24a1.61 1.61 0 011.61 1.61v3.69h4.12v-3.69a5.74 5.74 0 00-5.73-5.73z'
+                            />
+                            <path className='christmas_tree__cls-3' d='M110.54 680.5v5.25H124.39v-5.25H110.54z' />
+                            <path
+                                className='christmas_tree__cls-12'
+                                d='M96.93 546.46a19.72 19.72 0 00-6.23 9c-.09.26-.16.53-.24.8.07 0 4.21 1 4.21 1a1.15 1.15 0 01.82 1.39 1.14 1.14 0 01-1.37.84l-4.11-1-2 14.7c-.07.31-.11.62-.17.93.64.22 12 4 19.75 3.83 6.53-.16 7.39-1.33 10.57-7.11l3.13-5.67a65.37 65.37 0 01-24.36-18.71zm-2.31 16.25c-.27 1.09-1.06 1.84-1.76 1.67s-1-1.2-.78-2.29 1-1.84 1.76-1.67 1.05 1.2.78 2.29zm6.37 10a1.77 1.77 0 01-2.63 1.51l-2-1.14a2.31 2.31 0 01-.7-3.39l.25-.33a2.26 2.26 0 001.09 2.2l2 1.14a1.7 1.7 0 002-.24zm9.08-5.85c-.27 1.09-1.06 1.83-1.76 1.66s-1-1.2-.78-2.29 1-1.83 1.75-1.66 1.05 1.17.79 2.26zm-4.6-5.72a1.14 1.14 0 011.37-.84l4.14 1a1.15 1.15 0 01.82 1.39 1.14 1.14 0 01-1.37.84l-4.14-1a1.14 1.14 0 01-.82-1.42zM126 571.3l-2.16 10.17c-.77 4.65-3 9.49-6.75 12.23a44.46 44.46 0 01-8.36 4.26c-1.85.65-4.65.44-7.5-.08l-1.88 9.59a7 7 0 007.1 8.26l9.14-.36a5.33 5.33 0 005-4.31l4-20.45 1.58-8.1a6.45 6.45 0 003.08-4.22c.75-3.2-.7-6.29-3.25-6.99z'
+                            />
+                            <path
+                                className='christmas_tree__cls-5'
+                                d='M111 561.28l-4.14-1a1.14 1.14 0 00-.55 2.22l4.14 1a1.14 1.14 0 001.37-.84 1.15 1.15 0 00-.82-1.38zM90.46 556.25l4.21 1a1.15 1.15 0 01.82 1.39 1.14 1.14 0 01-1.37.84l-4.11-1a1.14 1.14 0 01.45-2.24z'
+                            />
+                            <path
+                                className='christmas_tree__cls-11'
+                                d='M94.62 562.71c-.27 1.09-1.06 1.84-1.76 1.67s-1-1.2-.78-2.29 1-1.84 1.76-1.67 1.05 1.2.78 2.29zM109.28 564.54c-.7-.17-1.48.57-1.75 1.66s.08 2.12.78 2.29 1.49-.57 1.76-1.66-.07-2.12-.79-2.29z'
+                            />
+                            <path
+                                className='christmas_tree__cls-16'
+                                d='M99 572.7l-2-1.14a2.26 2.26 0 01-1.11-2.23l-.25.33a2.31 2.31 0 00.7 3.39l2 1.14a1.77 1.77 0 002.63-1.51v-.22a1.7 1.7 0 01-1.97.24z'
+                            />
+                            <path
+                                className='christmas_tree__cls-5'
+                                d='M155.87 537.67a6.54 6.54 0 107.69-5.11 6.52 6.52 0 00-7.69 5.11z'
+                            />
+                            <path
+                                className='christmas_tree__cls-6'
+                                d='M146.28 553.58c-.24 6.73-5.55 21.08-9.83 26.44L98 539.58a2 2 0 01.49-3.21c10.14-5.43 49-6 57.5-2.14 10.93 4.95-9.42 10.83-9.71 19.35z'
+                            />
+                            <path
+                                className='christmas_tree__cls-5'
+                                d='M137.3 576.73a15.5 15.5 0 00-3-5.73l-31.13-31.06a6.48 6.48 0 00-10.39 1.52c-.22.42-.44.88-.65 1.4a49.75 49.75 0 00-2.77 8.94c-.23 1.11-.46 1.92.29 2.89a2.47 2.47 0 001 .77 19.72 19.72 0 016.23-9 65.37 65.37 0 0024.41 18.71l-3.13 5.67c-3.18 5.78-4 6.95-10.57 7.11-7.76.18-19.11-3.61-19.75-3.83-.07 0-2.13 12.27-2.13 12.27-.17 3.68 2.65 8.27 6.18 9.2 2.08.54 5.71 1.65 9.24 2.29 2.85.52 5.65.73 7.5.08a44.46 44.46 0 008.37-4.26c3.8-2.74 6-7.58 6.75-12.23L126 571.3c2.6.69 4.08 3.79 3.31 7a6.45 6.45 0 01-3.08 4.22l-1.58 8.1c1-.35 2 0 3-.25 2.49-.67 2.44-4.46 4.43-6.12 1.34-1.11 3.42-1 4.64-2.3s1.08-3.44.6-5.21zm-39.57 6.59s-3.21-.55-3.28-3.3l7.76 1.1c-.43 1.18-1.49 2.48-4.48 2.2z'
+                            />
+                            <path
+                                className='christmas_tree__cls-11'
+                                d='M97.73 583.32c3 .28 4-1 4.48-2.2L94.45 580c.07 2.77 3.28 3.32 3.28 3.32z'
+                            />
+                            <path
+                                className='christmas_tree__cls-5'
+                                d='M95.86 540.3a148.79 148.79 0 0123.79 17.29c6.46 5.55 13.75 11.3 16.93 19.48 1.9 4.88 9.86 2.76 7.94-2.19-3.56-9.17-10.39-15.62-17.69-21.94-8.41-7.28-17.09-14.28-26.82-19.75-4.62-2.6-8.78 4.51-4.15 7.11zM18.19 619.52a42.88 42.88 0 00-5.5 10.19c-.8 2.08.89 4.52 2.88 5.07 2.33.64 4.26-.8 5.06-2.88a35.53 35.53 0 014.67-8.22c1.33-1.79.3-4.6-1.47-5.63-2.1-1.23-4.3-.32-5.64 1.47zM148.86 671.22c-.08 6.49.47 13.36 5.65 17.94 4 3.5 9.8-2.3 5.82-5.82-3.48-3.09-3.28-7.81-3.23-12.12.07-5.3-8.17-5.3-8.24 0zM65.9 717.75q24.25 1.46 48.52 2.91c15.15.89 30.34 1.32 45.17-2.43 6.42-1.63 3.7-11.56-2.74-9.93-14.51 3.68-29.45 2.84-44.26 2l-46.69-2.8c-6.62-.4-6.6 9.9 0 10.29zM83.51 591.83c.3 6.63 2.72 12.54 1.94 19.22a56.48 56.48 0 00-.45 10.48 22.86 22.86 0 00.6 4.52l.5 1.35c.64 2 .81 2 .5.06a6 6 0 001.59 6c4.84 4.27 10.32 1 13.41-3.62 1.24-1.88 2-4 3.59-5.47 2.12-2 3.64-2.66 5.07-5.4 2.78-5.28 6.84-10.19 8.63-16 1.62-5.19 2.76-9.42 2.51-14.95-.25-5.74-8.59-8.9-11.51-3.12-2.53 5-4.69 11.08-8.67 15.14s-7.84 7.37-10.07 12.96l11.91 3.28c1.43-6.2 2.69-12.43 2.16-18.83-.51-6.11-8-8.39-11.51-3.11l-.37.54L97 592c1.95-.68 1 .19-.59-1.1a3.67 3.67 0 01-1.41-2.14c-.84-2.4-.3-5.8-.38-8.31-.24-7.93-12.59-8-12.35 0 .21 7-.16 14.25 5.47 19.31 4.68 4.21 12.54 7 16.67.8l-11.51-3.12c.45 5.37-.51 10.34-1.71 15.55-1.83 7.92 9 10.53 11.91 3.28 1.73-4.34 7.22-7.25 10-11 3.09-4.28 5.19-9.34 7.56-14l-11.51-3.11c0 .91.47 3.78.41 4.12-.29 1.5-1.44 3.2-1.85 4.78-1.11 4.27-3.16 7.31-5.46 11-1 1.6-1.66 3.79-2.78 5.24-.72.94-2.34 1.64-3.31 2.58a20 20 0 00-3.21 4.36c-.4.68-2.12 3.39-1.55 3.28l6 1.59-.38-.34 1.58 6c1.22-3-.3-5.2-1-8-.86-3.4-.21-7.48.16-10.91a35.23 35.23 0 000-9.89c-.56-3.4-1.68-6.68-1.84-10.14-.36-7.91-12.72-8-12.35 0z'
+                            />
+                            <path
+                                className='christmas_tree__cls-5'
+                                d='M93.58 572.49a9.89 9.89 0 00-5.47-.43 17.69 17.69 0 01-6.84-.09 1.66 1.66 0 00-.88 3.2c3.54 1.29 8 3.72 11.89 3.1 2.76-.44 4.23-3.51 1.66-5.31-1.75-1.23-3.41 1.64-1.67 2.86 0-.81-.27-1.07-1-.8a8.78 8.78 0 01-1.84-.26 22.74 22.74 0 01-2.53-.76c-1.89-.63-3.76-1.36-5.63-2l-.88 3.2a17.3 17.3 0 005.64.54c2.3-.24 4.42-1 6.67 0 2 .82 2.82-2.38.88-3.19zM104 577.62c.39.52 1.88 1 2.49 1.33s1.27.8 1.94 1.14a21.89 21.89 0 004.83 1.72l.9-3.26c-1.56-.3-3.12-.17-4.68-.46s-3.11-.71-4.66-1.06c-2.12-.48-3 2.78-.9 3.27 1.55.35 3.1.74 4.66 1.05s3.14.17 4.68.46c2.12.4 3-2.77.9-3.26a19.63 19.63 0 01-6-2.53 8 8 0 00-4.12-1.6c-2.49 0-4.73 2.63-3 4.91 1.29 1.72 4.24 0 2.92-1.71zM107.22 660a2.71 2.71 0 103.19-2.12 2.71 2.71 0 00-3.19 2.12zM107.22 646.75a2.71 2.71 0 103.19-2.12 2.7 2.7 0 00-3.19 2.12zM107.22 631.61a2.71 2.71 0 103.19-2.11 2.7 2.7 0 00-3.19 2.11z'
+                            />
+                            <g id='christmas_tree__bell'>
+                                <path
+                                    className='christmas_tree__cls-22'
+                                    transform='rotate(10.61 6.057 619.132)'
+                                    d='M3.95 605.67h4.19v26.67H3.95z'
+                                />
+                                <path
+                                    className='christmas_tree__cls-3'
+                                    d='M8.38 606.52a6.44 6.44 0 01-5.23-6.88l.19-12.49 16.77 3.14L15.75 602a6.43 6.43 0 01-7.37 4.52z'
+                                />
+                                <rect
+                                    className='christmas_tree__cls-3'
+                                    x={0.06}
+                                    y={585.64}
+                                    width={23.33}
+                                    height={6.16}
+                                    rx={3.08}
+                                    transform='rotate(10.61 11.717 588.844)'
+                                />
+                                <path
+                                    className='christmas_tree__cls-4'
+                                    d='M3.23 594.45l14.33 2.68-.99 2.67-13.38-2.5.04-2.85z'
+                                />
+                                <circle className='christmas_tree__cls-3' cx={12.43} cy={584.96} r={4.94} />
+                                <path
+                                    className='christmas_tree__cls-4'
+                                    transform='rotate(10.61 12.327 585.474)'
+                                    d='M6.44 585.01h11.81v.69H6.44z'
+                                />
+                            </g>
+                        </g>
+                    </g>
+                </g>
+            </svg>
+        </StyledRoom>
+    );
+};
+
+export default memo(ChristmasTree);
