+++ src/modules/events/views/ui/layouts/types/onsite/christmas-onsite-2020/rooms/InfoSignpost.tsx
@@ -1,0 +1,77 @@
+import { memo } from "react";
+import styled from "styled-components";
+
+import Room from "./Room";
+
+const StyledRoom = styled(Room)`
+    width: 8%;
+    position: absolute;
+    left: 68%;
+    top: 16vw;
+    z-index: 9;
+`;
+
+// eslint-disable-next-line @typescript-eslint/naming-convention
+interface InfoSignpostProps extends React.ComponentPropsWithoutRef<typeof Room> {}
+
+const InfoSignpost: React.FC<InfoSignpostProps> = (props) => {
+    return (
+        <StyledRoom {...props}>
+            <svg viewBox='0 0 384.41 603.84'>
+                <defs>
+                    <style>
+                        {".snowman__cls-1{fill:#994925}.snowman__cls-2{fill:#141f49}.snowman__cls-5{fill:#fff}"}
+                    </style>
+                </defs>
+                <g id='snowman__Layer_2' data-name='Layer 2'>
+                    <g id='snowman___Layer_' data-name='&lt;Layer&gt;'>
+                        <g id='snowman__info_signpost' data-name='info signpost'>
+                            <path className='snowman__cls-1' d='M241.45 474.99h-23.99V7.66L241.45 0v474.99z' />
+                            <path className='snowman__cls-2' d='M241.45.93h2.28v471.4h-2.28z' />
+                            <path
+                                fill='#c05e2d'
+                                d='M384.42 128.65l-301.54 9.44 1.69-37.84 16.99-12.08-16.17-6.13 1.25-27.94 295.15-9.23 1.46 46.61-38.65 8.77 39.17 7.66.65 20.74z'
+                            />
+                            <path
+                                d='M213.21 459.57c6.48 5.87 13 12.68 20.52 13.22 2.81.21 5.62-.5 8.43-.32a9 9 0 017.6 4.79c.11.21.22.43.32.66h-36.87z'
+                                fill='#caeafc'
+                            />
+                            <path
+                                className='snowman__cls-5'
+                                d='M127.47 83.1a2 2 0 01.64 1.45l.44 19.62a2 2 0 01-.58 1.49 2 2 0 01-1.45.64h-.36a2 2 0 01-1.49-.58 2 2 0 01-.64-1.45l-.44-19.62a2 2 0 01.58-1.49 2 2 0 011.45-.64h.38a2 2 0 011.47.58zM149.42 82.61a2 2 0 01.64 1.45l.44 19.62a2 2 0 01-.58 1.49 2 2 0 01-1.45.64h-.61a2 2 0 01-1.82-1l-7.75-12.53.27 11.66a2 2 0 01-.58 1.48 2 2 0 01-1.46.65h-.36a2 2 0 01-1.48-.58 2 2 0 01-.64-1.46l-.44-19.62a2 2 0 01.57-1.48 2 2 0 011.46-.65h.61a1.92 1.92 0 011.82 1l7.74 12.53-.26-11.66a2 2 0 01.58-1.48 2 2 0 011.45-.64h.36a2 2 0 011.49.58zM168.75 82.17a2 2 0 01.64 1.46V84a2 2 0 01-.58 1.48 2 2 0 01-1.45.64l-7.1.16.11 5 4.41-.09a2 2 0 011.48.57 2 2 0 01.65 1.46v.36a2 2 0 01-.57 1.48 2 2 0 01-1.46.65l-4.41.09.17 7.64a2 2 0 01-.58 1.49 2 2 0 01-1.45.64h-.36a2 2 0 01-1.49-.58 2 2 0 01-.64-1.45l-.44-19.62a2.08 2.08 0 012-2.13l9.55-.22a2 2 0 011.52.6zM188.83 87.82a21.56 21.56 0 01.65 5.19 21.15 21.15 0 01-.42 5.18 10.45 10.45 0 01-1.67 3.84 7 7 0 01-2.92 2.47 9.89 9.89 0 01-7.21.16 7.84 7.84 0 01-3.07-2.33 11 11 0 01-1.81-3.77 21.28 21.28 0 01-.65-5.15 21.82 21.82 0 01.42-5.21 10.89 10.89 0 011.67-3.85 7.65 7.65 0 012.93-2.47 9.73 9.73 0 013.6-.76 9 9 0 013.61.6 6.9 6.9 0 013.04 2.34 10.63 10.63 0 011.83 3.76zM185 93.11a16.47 16.47 0 00-.49-4.08 7.7 7.7 0 00-.84-2 3.92 3.92 0 00-1-.93 3.71 3.71 0 00-2.16-.44 3.84 3.84 0 00-2.18.53 3.59 3.59 0 00-.95 1 6.81 6.81 0 00-.78 2.06 16.37 16.37 0 00-.3 4.1 15.76 15.76 0 00.49 4.06 6.72 6.72 0 00.87 2 3.46 3.46 0 001 .93 3.87 3.87 0 002.19.43 3.82 3.82 0 002.15-.53 4.28 4.28 0 001-1 8.16 8.16 0 00.74-2 16.07 16.07 0 00.26-4.13zM207.26 102.37a1.71 1.71 0 01-.21 1 1.9 1.9 0 01-1.77 1.11h-.54a2.54 2.54 0 01-1.06-.23 1.34 1.34 0 01-.72-.74l-3.28-7.43h-1.01l.14 6.42a2 2 0 01-.57 1.48 2 2 0 01-1.46.65h-.36a2.07 2.07 0 01-2.12-2l-.45-19.62a2 2 0 01.58-1.48 2 2 0 011.46-.65l3.12-.07a10.37 10.37 0 012.77.3 6.16 6.16 0 012.68 1.3 6.66 6.66 0 011.82 2.59 8.78 8.78 0 01.72 3.5 8.48 8.48 0 01-.36 2.88 7.08 7.08 0 01-1.17 2.25 6.37 6.37 0 01-1.3 1.18l2.8 6.54a3.88 3.88 0 01.29 1.02zm-8.69-10.71h.54a4.44 4.44 0 002.74-.75 1.66 1.66 0 00.54-1.05 7.37 7.37 0 00.09-1.29 7.1 7.1 0 00-.17-1.36 2.08 2.08 0 00-.6-1.14 4.08 4.08 0 00-2.71-.67h-.61zM231.27 80.77a2 2 0 01.64 1.46l.44 19.62a2 2 0 01-.58 1.48 2 2 0 01-1.45.64H230a2.08 2.08 0 01-2.13-2l-.21-9.47-3.1 7.46a2 2 0 01-.75.93 2 2 0 01-1.12.37h-.43a2.09 2.09 0 01-1.93-1.22L217 92.86l.21 9.33a2 2 0 01-.58 1.48 2 2 0 01-1.46.64h-.35a2 2 0 01-1.49-.58 2 2 0 01-.64-1.45l-.44-19.62a2 2 0 01.58-1.49 2 2 0 011.45-.64h.61a2 2 0 011.14.31 2 2 0 01.79.9l5.48 11.86 5-12.1a2.13 2.13 0 01.75-.93 2.07 2.07 0 011.12-.36h.61a2 2 0 011.49.56zM254.65 102.57a2 2 0 01-1.66.89h-.4a2.06 2.06 0 01-1.23-.36 2.13 2.13 0 01-.77-1l-1.41-3.74-6.07.14-1.11 3.78a2.06 2.06 0 01-.73 1 2 2 0 01-1.21.42h-.36a2 2 0 01-1.67-.82 2 2 0 01-.36-1.86l5.87-19.8a1.81 1.81 0 01.71-1 2.08 2.08 0 011.23-.43h.46a2 2 0 011.23.37 2.07 2.07 0 01.78 1l7 19.51a2 2 0 01-.3 1.9zm-7.08-8.67l-1.65-4.85-1.47 4.95zM272.66 79.84a2 2 0 01.64 1.46v.36a2 2 0 01-.58 1.48 2 2 0 01-1.45.64l-3.41.08.39 17.18a2 2 0 01-.58 1.48 2 2 0 01-1.46.65h-.36a2 2 0 01-1.48-.58 2 2 0 01-.64-1.46L263.35 84H260a2.08 2.08 0 01-2.13-2v-.36a2 2 0 01.58-1.48 2 2 0 011.46-.64l11.3-.26a2 2 0 011.45.58zM281.63 79.64a2 2 0 01.64 1.46l.44 19.62a2 2 0 01-.58 1.48 2 2 0 01-1.45.64h-.36a2 2 0 01-1.49-.58 2 2 0 01-.64-1.45l-.44-19.62a2.08 2.08 0 012-2.13h.36a2 2 0 011.52.58zM304.4 85.23a22.08 22.08 0 01.65 5.19 21.06 21.06 0 01-.42 5.17 10.43 10.43 0 01-1.63 3.85 7 7 0 01-2.93 2.47 9.89 9.89 0 01-7.21.16 7.76 7.76 0 01-3.06-2.34A10.63 10.63 0 01288 96a25.37 25.37 0 01-.23-10.37 10.66 10.66 0 011.67-3.84 7.54 7.54 0 012.92-2.47 9.63 9.63 0 013.61-.76 9 9 0 013.6.6 6.84 6.84 0 013.07 2.34 10.89 10.89 0 011.76 3.73zm-3.87 5.29a16.55 16.55 0 00-.48-4.08 8.17 8.17 0 00-.84-2 4.18 4.18 0 00-1-.93 3.8 3.8 0 00-2.16-.43 3.88 3.88 0 00-2.18.53 3.51 3.51 0 00-.94 1 7 7 0 00-.78 2.07 16.33 16.33 0 00-.3 4.1 16.3 16.3 0 00.48 4.06 6.78 6.78 0 00.87 2 3.45 3.45 0 001 .93 5.09 5.09 0 004.34-.09 3.93 3.93 0 001-1 7.49 7.49 0 00.75-2 16.15 16.15 0 00.24-4.16zM325.24 78.66a2 2 0 01.64 1.46l.45 19.62a2 2 0 01-.58 1.48 2 2 0 01-1.46.64h-.61a2 2 0 01-1.82-1l-7.74-12.53.26 11.66a2 2 0 01-.58 1.48 2 2 0 01-1.45.64H312a2 2 0 01-1.49-.58 2 2 0 01-.64-1.45l-.44-19.62a2.08 2.08 0 012-2.13h.61a1.9 1.9 0 011.82 1l7.75 12.53-.26-11.66a2 2 0 01.57-1.49 2.05 2.05 0 011.46-.64h.36a2 2 0 011.5.59z'
+                            />
+                            <path
+                                className='snowman__cls-1'
+                                d='M115.73 123.59c28.27-1.44 56.57-2.2 84.86-3.21.64 0 .64-1 0-1-28.29 1-56.59 1.77-84.86 3.21-.64 0-.65 1 0 1zM227.82 118.37a121.38 121.38 0 0015.13-.73c4.89-.55 9.69-1.68 14.57-2.24 10.24-1.18 20.9.26 31.16.59q31.56 1 63.13 0c.64 0 .65-1 0-1q-31.55 1-63.13 0c-10.26-.33-20.92-1.77-31.16-.59-4.88.56-9.68 1.69-14.57 2.24a121.38 121.38 0 01-15.13.73.5.5 0 000 1zM150.66 66.73c10.78-.42 21.54-3.05 32.33-1.31 4.57.73 9 1.94 13.69 2.13a89.59 89.59 0 0015.6-1.21c.64-.09.37-1-.26-1-5.4.73-10.85 1.47-16.31 1.13-4.86-.31-9.55-1.72-14.38-2.33-10.23-1.28-20.45 1.17-30.67 1.56-.64 0-.64 1 0 1zM99.46 66.55l31.92.14a.5.5 0 000-1l-31.92-.14a.5.5 0 000 1zM257.34 65.69c17.05-2.63 34.23-5.27 51.54-4.94 16.6.32 32.76 5.28 49.37 5.06a.5.5 0 000-1c-16.93.22-33.41-4.87-50.35-5.08s-34 2.4-50.83 5c-.63.1-.37 1.07.27 1z'
+                            />
+                            <g id='snowman__snowman'>
+                                <path
+                                    className='snowman__cls-5'
+                                    d='M20.34 482.28C11 490.62 3.47 501.4.9 513.63c-4.09 19.42 6.1 40.94 23.75 50.11 7.51 3.91 16 5.7 23.44 9.7 8.09 4.35 14.58 11.09 22.06 16.4 15.25 10.84 34.63 15.48 53.27 13.59s36.43-10.14 50.5-22.48c11.94-10.48 21.43-24.23 24.74-39.75s-.17-32.8-10.92-44.5c-5.84-6.34-13.45-10.8-19.58-16.87-6.51-6.45-11.2-14.56-17.82-20.9-12.67-12.11-31.41-16.41-48.86-14.62-27.82 2.84-60.41 19.42-81.14 37.97z'
+                                />
+                                <path
+                                    d='M161.21 399.61c6.92 6.65 12.43 15.11 14.07 24.56 2.61 15-5.79 31.23-19.57 37.79-5.86 2.8-12.41 4-18.23 6.82-6.31 3.11-11.48 8.11-17.35 12a60.57 60.57 0 01-41.22 9 70.23 70.23 0 01-38.12-18.58C31.91 462.82 25 452 22.89 440s1-25.15 9.57-33.83c4.64-4.71 10.6-7.92 15.46-12.41 5.17-4.77 9-10.87 14.23-15.55 10-8.95 24.52-11.75 37.86-9.91 21.26 2.95 45.8 16.53 61.2 31.31z'
+                                    fill='#eef3ff'
+                                />
+                                <path
+                                    className='snowman__cls-5'
+                                    d='M65.49 322.86c-4.78 4.28-8.65 9.8-10 16.07a24.2 24.2 0 0012.17 25.68c3.85 2 8.2 2.91 12 5 4.15 2.22 7.47 5.68 11.3 8.4a40.49 40.49 0 0027.3 7 46.84 46.84 0 0025.87-11.52c6.12-5.37 11-12.41 12.68-20.36s-.09-16.81-5.6-22.8c-3-3.25-6.89-5.54-10-8.65-3.33-3.3-5.74-7.46-9.13-10.71-6.49-6.2-16.09-8.4-25-7.49-14.28 1.39-30.98 9.88-41.59 19.38z'
+                                />
+                                <path
+                                    d='M119.76 348.37a7.22 7.22 0 007.59 7.45l54.89-2.76a.88.88 0 00.12-1.74l-54.09-9.79a7.22 7.22 0 00-8.51 6.84z'
+                                    fill='#ee6d55'
+                                />
+                                <path
+                                    className='snowman__cls-2'
+                                    d='M97.33 338.89c6.44 0 6.45-10 0-10s-6.44 10 0 10zM131.48 334.45c6.43 0 6.44-10 0-10s-6.45 10 0 10zM97.33 369.2c3.22 0 3.23-5 0-5s-3.22 5 0 5zM105.76 371.86c3.22 0 3.22-5 0-5s-3.22 5 0 5zM116 373.19c3.21 0 3.22-5 0-5s-3.23 5 0 5zM125.27 373.19c3.22 0 3.22-5 0-5s-3.22 5 0 5zM134.58 370.08c3.22 0 3.22-5 0-5s-3.22 5 0 5zM143.45 365.65c3.21 0 3.22-5 0-5s-3.23 5 0 5zM117.29 417c7.07 0 7.09-11 0-11s-7.09 11 0 11zM119.5 454.67c7.08 0 7.09-11 0-11s-7.09 11 0 11zM123.49 508.76c7.08 0 7.09-11 0-11s-7.09 11 0 11zM125.71 548.67c7.08 0 7.09-11 0-11s-7.09 11 0 11z'
+                                />
+                            </g>
+                        </g>
+                    </g>
+                </g>
+            </svg>
+        </StyledRoom>
+    );
+};
+
+export default memo(InfoSignpost);
