+++ src/modules/events/views/ui/layouts/types/onsite/christmas-onsite-2020/rooms/PuppetShow.tsx
@@ -1,0 +1,259 @@
+import { memo } from "react";
+import styled from "styled-components";
+
+import Room from "./Room";
+
+const StyledRoom = styled(Room)`
+    width: 9%;
+    position: absolute;
+    left: 3%;
+    top: -1vw;
+    z-index: 5;
+`;
+
+interface PuppetShowProps extends React.ComponentPropsWithoutRef<typeof Room> {}
+
+const PuppetShow: React.FC<PuppetShowProps> = (props) => {
+    return (
+        <StyledRoom {...props}>
+            <svg viewBox='0 0 360.01 328.34'>
+                <defs>
+                    <style>
+                        {
+                            ".puppet_show__cls-3{fill:#fff}.puppet_show__cls-6{fill:#ee6d55}.puppet_show__cls-7{fill:#141f49}.puppet_show__cls-8{fill:#b54637}.puppet_show__cls-9{fill:#d15447}.puppet_show__cls-10,.puppet_show__cls-19,.puppet_show__cls-20,.puppet_show__cls-23,.puppet_show__cls-26{fill:none;stroke-miterlimit:10}.puppet_show__cls-10{stroke:#d15447;stroke-width:.75px}.puppet_show__cls-11{fill:#d8402b}.puppet_show__cls-12{fill:#fcf8cd}.puppet_show__cls-13{fill:#ffe09c}.puppet_show__cls-14{fill:#d9f2fc}.puppet_show__cls-16{fill:#ff9985}.puppet_show__cls-17{fill:#c05e2d}.puppet_show__cls-18{fill:#994925}.puppet_show__cls-19,.puppet_show__cls-20,.puppet_show__cls-23,.puppet_show__cls-26{stroke:#141f49;stroke-linecap:round}.puppet_show__cls-19{stroke-width:.56px}.puppet_show__cls-20{stroke-width:.71px}.puppet_show__cls-21{fill:#fde7df}.puppet_show__cls-23{stroke-width:.52px}.puppet_show__cls-24{fill:#a9cced}.puppet_show__cls-25{fill:#6195c9}.puppet_show__cls-26{stroke-width:.66px}"
+                        }
+                    </style>
+                </defs>
+                <g
+                    style={{
+                        isolation: "isolate",
+                    }}
+                >
+                    <g id='puppet_show__Layer_2' data-name='Layer 2'>
+                        <g id='puppet_show___Layer_' data-name='&lt;Layer&gt;'>
+                            <g id='puppet_show__puppet_show' data-name='puppet show'>
+                                <path fill='#ffb072' d='M25.06 211.9h318.7v115.69H25.06z' />
+                                <path className='puppet_show__cls-3' d='M20.13 205.25h327.84v6.65H20.13z' />
+                                <path fill='#370625' d='M25.09 211.9h318.32v13.69H25.09z' opacity={0.26} />
+                                <path
+                                    className='puppet_show__cls-6'
+                                    d='M25.1 211.83h318.73v115.72H25.1zM326.24 49.25h8.5v156.06h-8.5zM30.51 49.21h8.5v156.06h-8.5z'
+                                />
+                                <path className='puppet_show__cls-7' d='M39.01 78.41h287.23v126.9H39.01z' />
+                                <path
+                                    className='puppet_show__cls-8'
+                                    d='M62 112.67c2.47 9.72 5.69 28.74-2 49.4a76.5 76.5 0 01-7.32 14.64 39.53 39.53 0 0113.64 28.54H38.97v-113l22.36 1z'
+                                />
+                                <path
+                                    className='puppet_show__cls-9'
+                                    d='M39 173.38h17.55a3.16 3.16 0 013.16 3.16 3.16 3.16 0 01-3.16 3.16H39v-6.32z'
+                                />
+                                <path
+                                    className='puppet_show__cls-10'
+                                    d='M56.6 118.84C56.6 146 54.85 168 52.68 168M47.6 114.84C47.6 142 45.85 164 43.68 164'
+                                />
+                                <path
+                                    className='puppet_show__cls-9'
+                                    d='M52.89 183.71a31.23 31.23 0 015.57 16.77c0 .48.77.49.75 0a32 32 0 00-5.67-17.14c-.27-.4-.92 0-.65.37zM44.52 184.44c.45 3 2 5.49 2.16 8.54.18 2.65.09 5.34.12 8a.38.38 0 00.75 0c0-3 .11-6-.19-8.95-.27-2.73-1.71-5.09-2.11-7.79-.07-.47-.8-.27-.73.2z'
+                                />
+                                <path
+                                    className='puppet_show__cls-8'
+                                    d='M303.22 112.67c-2.47 9.72-5.69 28.74 2 49.4a76.5 76.5 0 007.32 14.64 39.53 39.53 0 00-13.64 28.54h27.35v-113l-22.36 1z'
+                                />
+                                <path
+                                    className='puppet_show__cls-9'
+                                    d='M305.54 173.38h17.53a3.16 3.16 0 013.16 3.16 3.16 3.16 0 01-3.16 3.16h-17.53v-6.32z'
+                                    transform='rotate(180 315.89 176.54)'
+                                />
+                                <path
+                                    className='puppet_show__cls-10'
+                                    d='M308.65 118.84c0 27.15 1.75 49.11 3.92 49.11M317.65 114.84c0 27.15 1.75 49.11 3.92 49.11'
+                                />
+                                <path
+                                    className='puppet_show__cls-9'
+                                    d='M311.71 183.34a31.91 31.91 0 00-5.71 17.14c0 .49.73.49.75 0a31.23 31.23 0 015.57-16.77c.27-.39-.38-.77-.65-.37zM320 184.24c-.4 2.7-1.84 5.07-2.12 7.79-.29 2.95-.15 6-.18 8.95a.38.38 0 00.75 0c0-2.66-.06-5.35.12-8 .2-3 1.71-5.58 2.15-8.54.07-.47-.65-.67-.72-.2z'
+                                />
+                                <path
+                                    className='puppet_show__cls-11'
+                                    d='M25.1 229.68h318.73v.97H25.1zM25.1 249.44h318.73v.97H25.1zM25.1 269.2h318.73v.97H25.1zM25.1 288.96h318.73v.97H25.1zM25.1 308.72h318.73v.97H25.1z'
+                                />
+                                <path
+                                    className='puppet_show__cls-3'
+                                    d='M4.63 45.68v51.07a15.92 15.92 0 0015.93 15.92h7.09a15.93 15.93 0 0015.94-15.92V45.68zM317.38 45.68v51.07a15.92 15.92 0 0015.94 15.92h7.08a15.92 15.92 0 0015.94-15.92V45.68z'
+                                />
+                                <path
+                                    className='puppet_show__cls-6'
+                                    d='M43.59 45.68v51.07a15.93 15.93 0 0015.94 15.92h7.08a15.92 15.92 0 0015.94-15.92V45.68zM278.42 45.68v51.07a15.92 15.92 0 0015.93 15.92h7.09a15.93 15.93 0 0015.94-15.92V45.68z'
+                                />
+                                <path
+                                    className='puppet_show__cls-3'
+                                    d='M82.55 45.68v51.07a15.93 15.93 0 0015.94 15.92h7.09a15.92 15.92 0 0015.93-15.92V45.68zM239.45 45.68v51.07a15.93 15.93 0 0015.94 15.92h7.09a15.93 15.93 0 0015.94-15.92V45.68z'
+                                />
+                                <path
+                                    className='puppet_show__cls-6'
+                                    d='M121.51 45.68v51.07a15.93 15.93 0 0015.94 15.92h7.09a15.93 15.93 0 0015.94-15.92V45.68zM200.49 45.68v51.07a15.93 15.93 0 0015.94 15.92h7.08a15.92 15.92 0 0015.94-15.92V45.68z'
+                                />
+                                <path
+                                    className='puppet_show__cls-3'
+                                    d='M160.48 45.68v51.07a15.92 15.92 0 0015.94 15.92h8.13a15.92 15.92 0 0015.94-15.92V45.68z'
+                                />
+                                <path
+                                    className='puppet_show__cls-7'
+                                    d='M137.14 90.89c-45.11 0-88.48-9.19-113.5-38l.94-.81C64.23 97.75 150.8 93.56 216.43 82c71.24-12.55 133.88-36.25 134.51-36.49l.44 1.16c-.62.24-63.36 24-134.71 36.55-24.74 4.36-52.45 7.67-79.53 7.67z'
+                                />
+                                <path
+                                    className='puppet_show__cls-12'
+                                    d='M289.64 62.61a3.32 3.32 0 103.6-3 3.35 3.35 0 00-3.6 3zM231.67 75.52a3.36 3.36 0 003 3.64 3.32 3.32 0 003.6-3 3.37 3.37 0 00-3-3.64 3.32 3.32 0 00-3.6 3zM156.23 86.63a3.36 3.36 0 003 3.64 3.32 3.32 0 003.6-3 3.37 3.37 0 00-3-3.64 3.32 3.32 0 00-3.6 3zM79.06 82.22a3.32 3.32 0 103.6-3 3.35 3.35 0 00-3.6 3z'
+                                />
+                                <path
+                                    className='puppet_show__cls-13'
+                                    d='M265.49 73.55a3.18 3.18 0 002.87 3.45 3.13 3.13 0 003.4-2.88 3.15 3.15 0 10-6.27-.57zM194.68 87.31a3.17 3.17 0 002.86 3.45 3.14 3.14 0 003.46-2.87 3.18 3.18 0 00-2.86-3.45 3.14 3.14 0 00-3.46 2.87zM119.13 91.61a3.18 3.18 0 002.87 3.45 3.14 3.14 0 003.41-2.88 3.18 3.18 0 00-2.86-3.45 3.14 3.14 0 00-3.42 2.88zM41.33 71.13a3.15 3.15 0 103.41-2.88 3.14 3.14 0 00-3.41 2.88z'
+                                />
+                                <path
+                                    className='puppet_show__cls-3'
+                                    d='M348.38 51.77c-2.52-.54-5-1.56-7.57-1.28-3.45.36-6.27 3-9.74 2.86-4.78-.2-8.56-5.58-13.06-4.24-3.22 1-3.79 4.55-5.08 7.26-2.26 4.76-10.6 7.48-14.3 3.37-2.23-2.47-1.8-6.3-4.45-8.45-4.22-3.44-10.93.57-16.7.13-3.54-.27-6.65-2.29-10.17-2.66-5.92-.63-11.27 3.43-17.22 3.38-6.38-.06-11.67-4.75-18-5.32-5.09-.45-10.09 1.84-15.19 1.38-4.94-.44-9.31-3.41-14.27-3.59-9.45-.37-15.37 9-24.4 11.33-7.5 1.94-16.47-2.08-19-8.52-1.9-4.88-11.76-2.7-17.33-.91s-11.23 4.4-17.13 3.56c-4.23-.6-8.64-2.93-12.47-1.34-1.55.64-2.69 1.84-4.18 2.56-3.74 1.79-8.46.16-12.07-1.87s-7.25-4.58-11.54-4.68C78.66 44.61 74.28 49 69.63 52s-12.73 4.3-15.42-.23c-.94-1.59-.93-3.58-2.3-4.92-2.74-2.67-7.75-.49-10.76 1.89s-7.49 5-10.75 2.81c-2.61-1.74-2.14-5.25-3.81-7.69s-5.24-3.44-8.46-3.18a28.18 28.18 0 00-9 2.82c0-2.69-4-3.93-7.1-4H.7c.86-3.34 5-7.76 13.47-7.62 3.67.07 7.44.79 11 .85 9.28.16 11.32-6 26.11-5.73 15.28.27 26.82 8.4 31.3 8.48 5 .09 10.85-3.26 17-3.16 9.72.17 15.09 8.56 26.65 8.76 13.29.23 19.87-8.31 39.35-8 35.57.61 63.75 5.45 80.72 6.63 5.67.39 9.89 1.7 14.06 1.77 7.05.12 13.81-3.62 20.51-3.34 5.52.23 7.28 2.31 17.78 1.75 19.75-1 27.86-3.31 36-3.17a35 35 0 0119 5.87c2.6 1.73 5.44 3.24 6.14 5.43a5.26 5.26 0 01.26 1.62 3.09 3.09 0 01-.49 1.53 21.37 21.37 0 01-11.18.6z'
+                                />
+                                <path
+                                    className='puppet_show__cls-14'
+                                    d='M.66 39.49h1.33c3.13.09 7.07 1.33 7.1 4a28.18 28.18 0 019-2.82c3.22-.26 6.81.76 8.46 3.18s1.2 6 3.81 7.69c3.26 2.17 7.74-.44 10.75-2.81s8-4.56 10.76-1.89c1.37 1.34 1.36 3.33 2.3 4.92C56.9 56.34 65 55.09 69.63 52s9-7.43 14.91-7.3c4.29.1 7.92 2.65 11.54 4.68s8.33 3.66 12.07 1.87c1.49-.72 2.63-1.92 4.18-2.56 3.83-1.59 8.24.74 12.47 1.34 5.9.84 11.56-1.77 17.13-3.56s15.43-4 17.33.91c2.5 6.44 11.47 10.46 19 8.52 9-2.35 15-11.7 24.4-11.33 5 .18 9.33 3.15 14.27 3.59 5.09.46 10.08-1.83 15.17-1.38 6.35.57 11.64 5.26 18 5.32 6 0 11.3-4 17.22-3.38 3.52.37 6.63 2.39 10.17 2.66 5.77.44 12.48-3.57 16.7-.13 2.65 2.15 2.22 6 4.45 8.45 3.7 4.11 12 1.39 14.3-3.37 1.29-2.71 1.86-6.3 5.08-7.26 4.5-1.34 8.28 4 13.06 4.24 3.47.15 6.29-2.5 9.74-2.86 2.56-.28 5.05.74 7.57 1.28a21.37 21.37 0 0011.14-.58C356.46 56.29 340 61.45 339 61.44c-2.78 0-5.71-1-8.3-1.08-6.2-.11-10 11.86-25.37 11.59-17.72-.3-29.71-17.23-35.1-17.33-6.37-.11-7.21 2.3-13.78 4.05-5.8 1.55-12.57.59-17.85.5-10.87-.19-12.1-5.82-25.67-6.06-18.6-.32-33.08 14.16-39.4 14-9.26-.16-13.94-14.34-24.6-14.52-9.46-.17-19.5 11.85-24.57 11.76-3-.05-5.64-3-8.09-3.08-2.63 0-4.86 2.52-7.8 2.47-5.91-.1-5-7.25-9.67-7.33-7.29-.13-24.27 20.25-38 20-6.55-.12-10.92-4.4-13.5-8.12a11.71 11.71 0 00-9.56-4.59c-9.05-.1-14.24-6.7-19.11-13.6C15.9 46.22.4 49.36.49 40.73a5.12 5.12 0 01.17-1.24z'
+                                />
+                                <path
+                                    className='puppet_show__cls-14'
+                                    d='M3.38 40.71l-2.82.52a.13.13 0 00-.09.13L.3 55.21c0 .16.19.18.21 0l1-7.82c0-.12.13-.16.18-.06l.65 1.14c0 .09.17 0 .19-.06l1-7.57c-.02-.07-.08-.14-.15-.13zM334.08 60.55l6.93-.77c.1 0 .21 0 .2.08L339 70a.07.07 0 01-.13 0l.28-6.78a.06.06 0 00-.12 0l-1 3a.23.23 0 01-.43 0l-1.45-4.06c-.11 0-.92 1-.92 1L334 60.63s0-.07.08-.08zM123.33 63.38l4.36-.88c.06 0 .13 0 .12.08l-1.54 10.11c-.07 0 .2-6.79.2-6.79l-.73 3c0 .21-.23.22-.27 0l-.85-4a6.22 6.22 0 00-.59 1c-.07 0-.76-2.39-.76-2.39s.01-.12.06-.13z'
+                                />
+                                <path
+                                    className='puppet_show__cls-3'
+                                    d='M43.38 13.14c6.5.63 13 2.57 19.48 1.68 6.94-.94 13.61-5.08 20.58-5.26 8.15-.21 15.9 5 24 6.6C125.88 19.87 143.55 4.85 162 1a48.2 48.2 0 0133.75 5.57c7.63 4.29 14.9 10.67 23 12 8.55 1.4 17.2-3 25.75-1.46 4.65.83 9.13 3.39 13.81 3.66 5.26.3 10.36-2.32 15.49-4.07 15.74-5.39 32.25-2.69 48.05 2.25a194 194 0 0125.35 10c4.45 2.1 8.38 8.15 11.66 18.28-39-6.7-81-10.8-119.82-2.18-5.44 1.21-10.93 2.57-16.42 2.08-6.53-.58-12.8-3.75-19.2-5.73-9.88-3.06-20.06-3.29-30.16-3.5-22.15-.47-31.39 5.17-53.53 6.54-8.7.54-30.33-4.74-39-4.74s-17.2-.8-25.78-1.59l-27.69-2.64C22.12 35-.52 48.54 0 39.68 1.49 15 37.7 12.58 43.38 13.14z'
+                                />
+                                <path
+                                    d='M23.19 328.34c.24-.4.5-.79.77-1.16 4.32-5.83 11.38-8.05 18.07-8.37s13.38.91 20.07.55c18.14-1 34-13.34 49.69-23.74s34.28-19.41 51.57-13.09c14.49 5.3 25 20.46 39.65 25.33 23.24 7.76 54-14.55 74.68 5.62 3.71 3.6 6.24 9.17 7.08 14.86z'
+                                    fill='#caeafc'
+                                />
+                                <g id='puppet_show__bear'>
+                                    <path className='puppet_show__cls-16' d='M102.76 193.79h17.12v11.52h-17.12z' />
+                                    <rect
+                                        className='puppet_show__cls-17'
+                                        x={122.74}
+                                        y={163.54}
+                                        width={8.35}
+                                        height={20.27}
+                                        rx={4.17}
+                                        transform='rotate(60 126.918 173.668)'
+                                    />
+                                    <rect
+                                        className='puppet_show__cls-17'
+                                        x={91.64}
+                                        y={163.54}
+                                        width={8.35}
+                                        height={20.27}
+                                        rx={4.17}
+                                        transform='rotate(120 95.812 173.67)'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-18'
+                                        d='M125.58 193.79H97.06l4.87-28.21h18.78c.46 2.69.93 5.38 1.39 8.08z'
+                                    />
+                                    <circle className='puppet_show__cls-18' cx={124.47} cy={137.29} r={7.7} />
+                                    <circle className='puppet_show__cls-13' cx={124.47} cy={137.29} r={4.98} />
+                                    <circle className='puppet_show__cls-18' cx={98.35} cy={137.29} r={7.7} />
+                                    <circle className='puppet_show__cls-13' cx={98.35} cy={137.29} r={4.98} />
+                                    <path
+                                        className='puppet_show__cls-17'
+                                        d='M122.2 138.29h-21.76a2.75 2.75 0 00-2.66 2.1l-2.36 9.92a14.57 14.57 0 007.49 16.27 17.84 17.84 0 008.41 2.08 17.81 17.81 0 008.41-2.08 14.57 14.57 0 007.49-16.27l-2.36-9.92a2.73 2.73 0 00-2.66-2.1z'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-13'
+                                        d='M111.32 173.79c-8.56 0-8.56 20-8.56 20h17.12s0-20-8.56-20zM111.72 151.32h-.8c-4.23 0-7.91 2.45-8.89 5.92-.88 3.12 4.66 9 9.29 9 4.71 0 10.17-5.93 9.29-9-.98-3.47-4.61-5.92-8.89-5.92z'
+                                    />
+                                    <circle className='puppet_show__cls-7' cx={111.32} cy={156.79} r={2.19} />
+                                    <path
+                                        className='puppet_show__cls-19'
+                                        d='M107.36 161.33a5.54 5.54 0 007.93 0M111.32 156.98V163'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-20'
+                                        d='M101.38 148a2.67 2.67 0 015.34 0M115.56 147.91a2.67 2.67 0 015.34 0'
+                                    />
+                                </g>
+                                <g id='puppet_show__pig'>
+                                    <path
+                                        className='puppet_show__cls-16'
+                                        d='M172.66 193.79h17.12v11.52h-17.12zM195.49 193.79H167l4.87-28.21h18.77c.47 2.69.93 5.38 1.4 8.08zM176.8 137.89l-10.63 10.62-1.49-1.51a19 19 0 01-5.38-12l-.25-3.06a1.17 1.17 0 011.22-1.21l3.17.26a18.7 18.7 0 0111.8 5.32z'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-21'
+                                        d='M174.06 138.11l-7.67 7.66-1.48-1.48a13.74 13.74 0 01-3.89-8.67l-.18-2.2a.85.85 0 01.88-.88l2.29.19a13.45 13.45 0 018.51 3.84z'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-16'
+                                        d='M185.33 137.89L196 148.51l1.48-1.49a19 19 0 005.39-12l.25-3.06a1.17 1.17 0 00-1.22-1.21l-3.17.26a18.68 18.68 0 00-11.8 5.32z'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-21'
+                                        d='M188.08 138.11l7.67 7.66 1.48-1.48a13.73 13.73 0 003.88-8.67l.19-2.2a.87.87 0 00-.89-.88l-2.28.19a13.48 13.48 0 00-8.52 3.84z'
+                                    />
+                                    <circle cx={181.23} cy={153.01} r={16.6} fill='#f9cdaf' />
+                                    <path
+                                        className='puppet_show__cls-21'
+                                        d='M181.23 173.79c-8.57 0-8.57 20-8.57 20h17.13s0-20-8.56-20z'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-23'
+                                        d='M171 148.55a2 2 0 013.93 0M187.12 148.51a2 2 0 113.92 0'
+                                    />
+                                    <circle className='puppet_show__cls-7' cx={181.23} cy={162.49} r={2.87} />
+                                    <circle className='puppet_show__cls-21' cx={181.23} cy={156.44} r={6.59} />
+                                    <circle className='puppet_show__cls-7' cx={178} cy={156.44} r={1.68} />
+                                    <circle className='puppet_show__cls-7' cx={184.42} cy={156.44} r={1.68} />
+                                </g>
+                                <g id='puppet_show__mouse'>
+                                    <path className='puppet_show__cls-16' d='M244.59 193.79h17.12v11.52h-17.12z' />
+                                    <path
+                                        className='puppet_show__cls-24'
+                                        d='M265.19 184.27a1.42 1.42 0 01-.18-2.82c1.86-.23 5.16-1 6.33-3.24a4.86 4.86 0 00.48-2.16v-.55c.18-3.89.26-5.67.29-9a1.42 1.42 0 011.42-1.4 1.42 1.42 0 011.4 1.43c0 3.37-.11 5.16-.29 9.09v.54a7.66 7.66 0 01-.79 3.34c-1.33 2.58-4.27 4.22-8.49 4.75z'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-25'
+                                        d='M267.42 193.79h-28.53l4.88-28.21h18.77c.47 2.69.93 5.38 1.4 8.08z'
+                                    />
+                                    <circle className='puppet_show__cls-25' cx={238.29} cy={143} r={10.15} />
+                                    <circle className='puppet_show__cls-14' cx={238.29} cy={143} r={7.92} />
+                                    <circle className='puppet_show__cls-25' cx={267.88} cy={143} r={10.15} />
+                                    <circle className='puppet_show__cls-14' cx={267.88} cy={143} r={7.92} />
+                                    <path
+                                        className='puppet_show__cls-24'
+                                        d='M266.87 165.58H239.4a2.56 2.56 0 01-2.43-3c.6-4.11 2.74-15.86 7.78-17.83a23.85 23.85 0 0116.81 0c5 2 7.18 13.72 7.78 17.83a2.56 2.56 0 01-2.43 3z'
+                                    />
+                                    <path
+                                        className='puppet_show__cls-14'
+                                        d='M253.15 173.79c-8.56 0-8.56 20-8.56 20h17.13s0-20-8.57-20zM253.16 161.41h-.62c-3.28 0-5.81-.08-6.57-2.12-.68-1.82 3.6-5.29 7.18-5.29 3.63 0 7.86 3.47 7.18 5.29-.75 2.04-3.9 2.12-7.17 2.12z'
+                                    />
+                                    <circle className='puppet_show__cls-7' cx={253.15} cy={154} r={1.68} />
+                                    <path className='puppet_show__cls-19' d='M249.19 157.59a5.52 5.52 0 007.93 0' />
+                                    <path
+                                        className='puppet_show__cls-26'
+                                        d='M246.45 152.15a1.53 1.53 0 113.05 0M256.44 152.12a1.53 1.53 0 113.06 0'
+                                    />
+                                </g>
+                                <path
+                                    className='puppet_show__cls-17'
+                                    d='M271.03 275.77l-187.46-5.04 2.42-23.42 10.97-6.88-9.81-4.39 1.79-17.29 183.49 4.94-.78 28.97-24.3 4.05 24.03 6.17-.35 12.89z'
+                                />
+                                <path
+                                    className='puppet_show__cls-18'
+                                    d='M104.5 262.92c17.59.13 35.18.68 52.77 1.08a.31.31 0 000-.62c-17.59-.4-35.18-.95-52.77-1.08a.31.31 0 000 .62zM174.25 263.73c6.29.43 12.28-.43 18.53-.77 6.4-.35 13 .92 19.32 1.5q19.54 1.77 39.17 2.29a.31.31 0 000-.62q-19.61-.53-39.17-2.29c-6.21-.57-12.59-1.77-18.84-1.52-6.41.26-12.56 1.23-19 .79-.4 0-.4.59 0 .62zM128.23 228.9c6.77.11 13.6-1.12 20.29.4a73 73 0 007.86 1.73 53.26 53.26 0 0010-.14c.4 0 .4-.65 0-.62a56.31 56.31 0 01-9.54.19c-3.4-.33-6.58-1.54-9.93-2.11-6.18-1-12.48 0-18.7-.07a.31.31 0 000 .62zM96.46 226.94l19.81 1.24c.4 0 .39-.6 0-.62l-19.81-1.25c-.4 0-.4.6 0 .63zM194.38 232.11c10.57-1 21.21-2 31.83-1.23 10.45.76 20.42 4.44 30.87 5 .4 0 .4-.6 0-.62-10.45-.51-20.43-4.2-30.87-5-10.62-.77-21.26.24-31.83 1.24-.39 0-.4.66 0 .62z'
+                                />
+                                <path
+                                    className='puppet_show__cls-3'
+                                    d='M114.59 240.28a6.43 6.43 0 01.27 2.19 5.77 5.77 0 01-.42 2.19 3.42 3.42 0 01-.5.88 4.75 4.75 0 01-.64.7 4.18 4.18 0 01-1.68.78 7.12 7.12 0 01-1.75.17h-.78l-.1 4.06a1.29 1.29 0 01-.41.93 1.31 1.31 0 01-1 .37h-.23a1.33 1.33 0 01-.94-.41 1.3 1.3 0 01-.36-1l.32-12.56a1.24 1.24 0 01.41-.93 1.29 1.29 0 011-.36l2.34.05a6.52 6.52 0 011.83.28 4.13 4.13 0 011.61.94 3.83 3.83 0 011.03 1.72zm-5.42 4h.83a4 4 0 001-.07 1.4 1.4 0 00.6-.28 2.21 2.21 0 00.45-1.55 2.42 2.42 0 00-.37-1.68 2.19 2.19 0 00-1.55-.45h-.78zM127.28 238.19a1.27 1.27 0 01.36.95l-.18 7.25c0 .49 0 1-.07 1.39a10.27 10.27 0 01-.16 1.2 5.61 5.61 0 01-.83 2.18 3.78 3.78 0 01-1.76 1.47 4.88 4.88 0 01-1.08.31 7 7 0 01-1.27.07 7.19 7.19 0 01-1.26-.13 5.44 5.44 0 01-1-.36 3.85 3.85 0 01-1.73-1.52 5.53 5.53 0 01-.73-2.22c0-.37-.06-.78-.08-1.21s0-.9 0-1.39l.18-7.25a1.29 1.29 0 01.41-.93 1.27 1.27 0 011-.37h.23a1.35 1.35 0 011.3 1.37l-.18 7.3a11.53 11.53 0 00.1 2.2 2.89 2.89 0 00.37 1.09.64.64 0 00.2.23l.24.15a2.69 2.69 0 00.47.14 4.25 4.25 0 00.61.05 3.77 3.77 0 00.62 0 3.38 3.38 0 00.49-.11 1.45 1.45 0 00.46-.36 1.36 1.36 0 00.21-.36 4.9 4.9 0 00.21-.69 6.82 6.82 0 00.15-1c0-.37.06-.77.07-1.22l.18-7.29a1.33 1.33 0 01.41-.94 1.31 1.31 0 011-.36h.23a1.3 1.3 0 01.83.36zM138.85 240.89a6.43 6.43 0 01.27 2.19 6 6 0 01-.43 2.19 3.46 3.46 0 01-.49.89 5.31 5.31 0 01-.65.69 4.22 4.22 0 01-1.67.79 7.7 7.7 0 01-1.75.16h-.78l-.1 4.06a1.27 1.27 0 01-.42.93 1.24 1.24 0 01-1 .37h-.23a1.27 1.27 0 01-.93-.41 1.29 1.29 0 01-.36-.95l.31-12.56a1.3 1.3 0 01.42-.93 1.27 1.27 0 01.95-.36l2.34.06a6.54 6.54 0 011.83.27 4 4 0 011.6.94 3.84 3.84 0 011.09 1.67zm-5.43 4h.78a4.08 4.08 0 001-.08 1.2 1.2 0 00.59-.27 2.24 2.24 0 00.46-1.55 2.44 2.44 0 00-.38-1.69 2.18 2.18 0 00-1.55-.45h-.78zM150.05 241.17a6.43 6.43 0 01.27 2.19 6.06 6.06 0 01-.43 2.2 3.57 3.57 0 01-.49.88 4.74 4.74 0 01-.65.69 4.22 4.22 0 01-1.67.79 7.18 7.18 0 01-1.75.16h-.78l-.1 4.06a1.27 1.27 0 01-.42.93 1.24 1.24 0 01-.95.37h-.23a1.27 1.27 0 01-.93-.42 1.3 1.3 0 01-.37-.95l.32-12.55a1.29 1.29 0 01.41-.93 1.28 1.28 0 011-.37l2.34.06a6.93 6.93 0 011.83.27 4.21 4.21 0 011.6.94 4 4 0 011 1.68zm-5.43 4h.78a4.08 4.08 0 001-.08 1.2 1.2 0 00.59-.27 2.25 2.25 0 00.45-1.55 2.41 2.41 0 00-.37-1.69 2.18 2.18 0 00-1.55-.45h-.78zM161 251.37a1.29 1.29 0 01.36.95v.23a1.27 1.27 0 01-.42.93 1.28 1.28 0 01-.95.37l-5.92-.15a1.27 1.27 0 01-.93-.42 1.28 1.28 0 01-.37-.95l.32-12.55a1.29 1.29 0 01.41-.93 1.31 1.31 0 011-.37l5.7.15a1.32 1.32 0 011.29 1.36v.23a1.27 1.27 0 01-.42.93 1.24 1.24 0 01-.95.37l-4.13-.11-.08 3.06 3.08.07a1.24 1.24 0 01.92.42 1.26 1.26 0 01.35.95v.23a1.31 1.31 0 01-1.34 1.3l-3.08-.08-.09 3.49 4.25.15a1.26 1.26 0 011 .37zM173.79 239.36a1.27 1.27 0 01.37.95v.23a1.33 1.33 0 01-1.36 1.3l-2.18-.06-.28 11a1.33 1.33 0 01-.41.93 1.3 1.3 0 01-.95.36h-.23a1.27 1.27 0 01-.93-.42 1.24 1.24 0 01-.37-1l.28-11-2.16-.05a1.27 1.27 0 01-.93-.42 1.28 1.28 0 01-.37-1v-.23a1.33 1.33 0 01.41-.93 1.3 1.3 0 01.95-.36l7.23.18a1.29 1.29 0 01.93.52zM192.74 247.56a4.39 4.39 0 01.71 2.68 4.7 4.7 0 01-.37 1.78 4.06 4.06 0 01-1 1.49 4.79 4.79 0 01-3.52 1.17 6.2 6.2 0 01-2.2-.42 7 7 0 01-1.72-.94 1.3 1.3 0 01-.32-1.82l.15-.2a1.38 1.38 0 01.86-.58 1.23 1.23 0 011 .23c.15.1.32.2.5.3a4.65 4.65 0 00.59.3 3.14 3.14 0 001.26.24 3 3 0 00.83-.09 2 2 0 00.58-.31 1.72 1.72 0 00.35-.45 2.27 2.27 0 00.14-.75 4.61 4.61 0 000-.52.77.77 0 00-.19-.45 1.78 1.78 0 00-.39-.35 4.47 4.47 0 00-.83-.42l-.91-.37a11 11 0 01-1.42-.72 5.18 5.18 0 01-1.07-.88 4.12 4.12 0 01-1-3 4.27 4.27 0 01.41-1.85 3.67 3.67 0 011-1.36 4.72 4.72 0 011.49-.78 5 5 0 011.59-.19 6.08 6.08 0 012.06.39 6 6 0 011.42.75 1.22 1.22 0 01.57.8 1.26 1.26 0 01-.15 1l-.12.2a1.4 1.4 0 01-.84.64 1.29 1.29 0 01-1-.18 3 3 0 00-.47-.25l-.48-.24a2.84 2.84 0 00-1-.18 1.72 1.72 0 00-1.21.33l-.19.25a1.65 1.65 0 00-.15.78 3.14 3.14 0 000 .45.77.77 0 00.06.26.72.72 0 00.2.38 2.61 2.61 0 00.43.35l.38.19.46.2 1.11.47a5.2 5.2 0 012.4 1.67zM205.81 240.17a1.29 1.29 0 01.36 1l-.31 12.55a1.27 1.27 0 01-.42.93 1.24 1.24 0 01-.95.37h-.25a1.2 1.2 0 01-.92-.41 1.33 1.33 0 01-.36-1l.13-5.09-4.09-.12-.13 5.1a1.32 1.32 0 01-1.37 1.29h-.22a1.33 1.33 0 01-.94-.41 1.31 1.31 0 01-.36-1l.32-12.55a1.22 1.22 0 01.41-.93 1.3 1.3 0 01.95-.37h.23a1.33 1.33 0 011.3 1.36l-.12 4.57 4.09.1.11-4.57a1.32 1.32 0 01.4-.93 1.26 1.26 0 01.94-.36h.26a1.29 1.29 0 01.94.47zM219.87 244.43a14.32 14.32 0 01.26 3.33 13.78 13.78 0 01-.43 3.3 6.58 6.58 0 01-1.19 2.4 4.33 4.33 0 01-1.94 1.49 6.33 6.33 0 01-4.61-.11 4.92 4.92 0 01-1.89-1.59 6.79 6.79 0 01-1-2.46 13.71 13.71 0 01-.26-3.31 13.4 13.4 0 01.43-3.32 6.87 6.87 0 011.18-2.41 4.71 4.71 0 012-1.48 6 6 0 012.32-.38 5.71 5.71 0 012.29.49 4.4 4.4 0 011.89 1.59 7 7 0 01.95 2.46zm-2.63 3.26a11 11 0 00-.19-2.62 4.89 4.89 0 00-.47-1.32 2.66 2.66 0 00-.63-.63 2.49 2.49 0 00-1.37-.34 2.45 2.45 0 00-1.41.27 2.28 2.28 0 00-.63.6 4.22 4.22 0 00-.56 1.29 10.21 10.21 0 00-.32 2.61 9.88 9.88 0 00.19 2.61 4.23 4.23 0 00.49 1.31 2.36 2.36 0 00.61.63 2.52 2.52 0 001.39.34 2.36 2.36 0 001.38-.27 2.46 2.46 0 00.66-.6 5.32 5.32 0 00.54-1.28 10.52 10.52 0 00.32-2.6zM239.84 241.17a1.39 1.39 0 01.22.55 1.12 1.12 0 01-.05.6l-3.76 12.46a1.22 1.22 0 01-.48.72 1.35 1.35 0 01-.83.24h-.34a1.35 1.35 0 01-.81-.28 1.29 1.29 0 01-.45-.74L231.4 247l-2.31 7.63a1.22 1.22 0 01-.48.72 1.35 1.35 0 01-.83.24h-.34a1.35 1.35 0 01-.81-.28 1.29 1.29 0 01-.45-.74l-3.13-12.64a1.22 1.22 0 01.26-1.14 1.24 1.24 0 011.07-.5h.25a1.3 1.3 0 01.81.3 1.26 1.26 0 01.45.72l1.9 7.74 2.33-7.65a1.19 1.19 0 01.49-.69 1.3 1.3 0 01.82-.24h.3a1.35 1.35 0 01.8.29 1.22 1.22 0 01.46.71l1.94 7.76 2.28-7.63a1.31 1.31 0 011.31-.95h.25a1.28 1.28 0 011.07.52z'
+                                />
+                            </g>
+                        </g>
+                    </g>
+                </g>
+            </svg>
+        </StyledRoom>
+    );
+};
+
+export default memo(PuppetShow);
